DCE-101: Move Evaluation Functions into separate files
