+++ src/engine/board/structures/piece.rs
@@ -24,7 +24,7 @@ pub const BLACK_QUEEN: Piece = BLACK | QUEEN;
 pub const BLACK_KING: Piece = BLACK | KING;
 
 pub const PIECES_WITHOUT_PAWN: [Piece; 5] = [KNIGHT, BISHOP, ROOK, QUEEN, KING];
-pub const PROMO_PIECES: [Piece; 4] = [KNIGHT, BISHOP, ROOK, QUEEN];
+pub const PIECES_WITHOUT_PAWN_KING: [Piece; 4] = [KNIGHT, BISHOP, ROOK, QUEEN];
 pub const PIECES: [Piece; 6] = [PAWN, KNIGHT, KING, BISHOP, ROOK, QUEEN];
 pub const CLR_PIECES: [Piece; 12] = [
     WHITE_PAWN,
