+++ src/engine/evaluation/evaluation.rs
@@ -1,59 +1,41 @@
-use std::array;
-
-use rand::rand_core::block;
-
 use crate::engine::board::structures::board::Board;
-use crate::engine::board::structures::castling;
-use crate::engine::board::structures::castling::CastlingRights;
-use crate::engine::board::structures::color;
 use crate::engine::board::structures::color::*;
-use crate::engine::board::structures::piece;
 use crate::engine::board::structures::piece::*;
 use crate::engine::board::structures::square::*;
-use crate::engine::evaluation::eval_defs;
-use crate::engine::evaluation::eval_defs::*;
-use crate::engine::misc::bit_pos_utility::get_bit_rank;
+use crate::engine::evaluation::common_eval::CommonEvalTrait;
+use crate::engine::evaluation::imbalance_eval::ImbalanceEvalTrait;
+use crate::engine::evaluation::init_eval::InitEvalTrait;
+use crate::engine::evaluation::king_eval::KingEvalTrait;
+use crate::engine::evaluation::material_eval::MaterialEvalTrait;
+use crate::engine::evaluation::mobility_eval::MobilityEvalTrait;
+use crate::engine::evaluation::passed_pawn_eval::PassedPawnEvalTrait;
+use crate::engine::evaluation::pawn_eval::PawnEvalTrait;
+use crate::engine::evaluation::piece_eval::PieceEvalTrait;
+use crate::engine::evaluation::piece_eval::OUTPOST_RANKS;
+use crate::engine::evaluation::psqt_eval::PSQTEvalTrait;
+use crate::engine::evaluation::space_eval::SpaceEvalTrait;
+use crate::engine::evaluation::tempo_eval::TempoEvalTrait;
+use crate::engine::evaluation::threats_eval::ThreatsEvalTrait;
+use crate::engine::evaluation::trace_eval::TraceEvalTrait;
 use crate::engine::misc::bitboard::Bitboard;
 use crate::engine::misc::bitboard::BitboardTrait;
 use crate::engine::misc::bitboard::Iterator;
 use crate::engine::misc::const_utility::FILE_BITBOARD;
-use crate::engine::misc::const_utility::RANK_BITBOARD;
 use crate::engine::misc::display::display_board::print_eval;
 use crate::engine::move_generator::bishop::get_bishop_mask;
-use crate::engine::move_generator::bishop::get_bishop_mv;
 use crate::engine::move_generator::bishop::has_bishop_pair;
-use crate::engine::move_generator::bishop::BLACK_SQUARES;
-use crate::engine::move_generator::bishop::WHITE_SQUARES;
 use crate::engine::move_generator::generated::between::BETWEEN_BB;
-use crate::engine::move_generator::generated::bishop::BISHOP_BASE;
-use crate::engine::move_generator::generated::bishop::BISHOP_MASKS;
-use crate::engine::move_generator::generated::king;
 use crate::engine::move_generator::generated::king::KING_RING;
-use crate::engine::move_generator::generated::knight;
 use crate::engine::move_generator::generated::pawn::FORWARD_SPANS_LR;
-use crate::engine::move_generator::generated::pawn::ISOLATED_PAWN_LOOKUP;
 use crate::engine::move_generator::generated::pawn::PAWN_3_BEHIND_MASKS;
-use crate::engine::move_generator::generated::pawn::PAWN_ATTACK_LOOKUP;
-use crate::engine::move_generator::generated::pawn::PAWN_FORWARD_SPANS;
-use crate::engine::move_generator::generated::pawn::PAWN_MOVE_LOOKUP;
 use crate::engine::move_generator::king::get_king_mask;
-use crate::engine::move_generator::king::get_king_mv;
-use crate::engine::move_generator::king::has_near_open_files;
 use crate::engine::move_generator::knight::get_knight_mask;
-use crate::engine::move_generator::knight::get_knight_mv;
-use crate::engine::move_generator::pawn::get_all_pawn_forward_mask;
 use crate::engine::move_generator::pawn::get_all_pawn_left_att_mask;
 use crate::engine::move_generator::pawn::get_all_pawn_right_att_mask;
 use crate::engine::move_generator::pawn::get_pawn_2_att;
 use crate::engine::move_generator::pawn::get_pawn_att_mask;
-use crate::engine::move_generator::pawn::is_blocked_pawn;
-use crate::engine::move_generator::pawn::is_isolated_pawn;
-use crate::engine::move_generator::pawn::is_passed_pawn;
 use crate::engine::move_generator::queen::get_queen_mask;
 use crate::engine::move_generator::rook::get_rook_mask;
-use crate::engine::move_generator::rook::get_rook_mv;
-use crate::engine::move_generator::rook::is_rook_on_open_file;
-use crate::engine::move_generator::rook::is_rook_on_semi_open_file;
 
 // The Numbers (Tapered Eval) for the evaluation are taken from -> STOCKFISH SF_9
 
@@ -99,9 +81,7 @@ impl Evaluation {
             mg_test: [[0; 64]; 2],
             eg_test: [[0; 64]; 2],
             vec_test: Vec::with_capacity(200),
-
             outpost: [0; 2],
-
             open_file: [u64::MAX; 2],
             pawn_att_span: [0; 2],
             king_ring: [0; 2],
@@ -110,7 +90,6 @@ impl Evaluation {
             attacked_by: [0; 14],
             defended_by: [0; 14],
             attacked_by_2: [0; 2],
-
             defended_by_2: [0; 2],
             king_att_weight: [0; 2],
             king_att_count: [0; 2],
@@ -156,470 +135,27 @@ impl Evaluation {
     }
 }
 
-pub trait EvaluationTrait {
-    // NOTE: Functions That Initialize the Evaluation Structure
-    fn init(&mut self);
-    fn king_init(&mut self);
-    fn pawn_init(&mut self);
-    fn piece_init(&mut self);
-    fn determine_phase(&mut self);
-    fn tapered(&mut self, value: (isize, isize)) -> isize;
-    fn insufficient_material(&self) -> bool;
-    fn front_sq(&mut self, sq: usize, clr: Color) -> usize;
-    fn back_sq(&mut self, sq: usize, clr: Color) -> usize;
-    fn pin_att(&mut self, from: usize, to: usize, piece: Piece) -> u64;
-
-    fn sum(
-        &mut self,
-        color: Color,
-        square: Option<usize>,
-        piece: Option<Piece>,
-        value: (isize, isize),
-    );
-
-    fn calculate_score(&mut self) -> isize;
-
-    // NOTE: TRACE [Debugging purposes]
-    fn trace(
-        &mut self,
-        color: Color,
-        square: Option<usize>,
-        piece: Option<Piece>,
-        value: (isize, isize),
-    );
-    fn print_trace_board(&mut self, name: &str);
-    fn print_trace_log(&mut self, name: &str);
-    fn print_trace_score(&mut self, name: &str);
-    fn reset_trace(&mut self);
-
+pub trait EvaluationTrait:
+    CommonEvalTrait
+    + InitEvalTrait
+    + TraceEvalTrait
+    + MaterialEvalTrait
+    + PSQTEvalTrait
+    + ImbalanceEvalTrait
+    + MobilityEvalTrait
+    + PassedPawnEvalTrait
+    + PawnEvalTrait
+    + PieceEvalTrait
+    + ThreatsEvalTrait
+    + KingEvalTrait
+    + SpaceEvalTrait
+    + TempoEvalTrait
+{
     // NOTE: Main Evaluation Function (It has 11 sub evaluations)
     fn evaluation(&mut self) -> isize;
-
-    // NOTE: 1. MATERIAL Evaluation
-    fn material_eval(&mut self, clr: Color);
-    fn non_pawn_material_eval(&mut self, clr: Color) -> isize;
-    fn piece_material(&mut self, piece: Piece) -> (isize, isize);
-
-    // NOTE: 2. PSQT Evaluation
-    fn psqt_eval(&mut self, clr: Color);
-    fn piece_psqt(&mut self, piece: Piece, sq: usize) -> (isize, isize);
-
-    // NOTE: 3. IMBALANCE Evaluation
-    fn imbalance(&mut self, clr: Color);
-    fn imb_piece_count(&mut self, num: usize, clr: Color) -> isize;
-
-    // NOTE: 4. PAWNS Evaluation FIXME:
-    fn pawns_eval(&mut self, clr: Color);
-    fn single_pawn_eval(&mut self, sq: usize, clr: Color);
-    fn isolated_pawn(&mut self, sq: usize, clr: Color) -> bool;
-    fn opposed_pawn(&mut self, sq: usize, clr: Color) -> bool;
-    fn phalanx_pawn(&mut self, sq: usize, clr: Color) -> bool;
-    fn supported_pawn(&mut self, sq: usize, clr: Color) -> isize;
-    fn backward_pawn(&mut self, sq: usize, clr: Color) -> bool;
-    fn doubled_pawn(&mut self, sq: usize, clr: Color) -> bool;
-    fn connected_pawn(&mut self, sq: usize, clr: Color) -> bool;
-    fn connected_bonus(&mut self, sq: usize, clr: Color) -> isize;
-    fn weak_unopposed_pawn(&mut self, sq: usize, clr: Color) -> bool;
-    fn weak_lever(&mut self, sq: usize, clr: Color) -> bool;
-    fn blocked_pawn(&mut self, sq: usize, clr: Color, bb: u64) -> bool;
-    fn blocked_pawn_5th_6th_rank(&mut self, sq: usize, clr: Color) -> isize;
-    fn doubled_isolated_pawn(&mut self, sq: usize, clr: Color) -> bool;
-
-    // NOTE: 5. PEACES Evaluation TODO: FIXME:
-    fn piece_eval(&mut self, clr: Color);
-    fn outpost(&mut self, clr: Color) -> u64;
-    fn reachable_outpost(&mut self, clr: Color) -> u64;
-    fn minor_behind_pawn(&mut self, clr: Color) -> isize;
-    fn bishop_pawns(&mut self, clr: Color) -> isize;
-    fn trapped_rook(&mut self, clr: Color);
-    fn weak_queen(&mut self, clr: Color) -> isize;
-    fn king_protector(&mut self, clr: Color);
-    fn outpost_total(&mut self, clr: Color);
-    fn rook_on_queen_file(&mut self, clr: Color);
-    fn bishop_xray_pawns(&mut self, clr: Color) -> isize;
-    fn rook_on_king_ring(&mut self, clr: Color) -> isize;
-    fn bishop_on_king_ring(&mut self, clr: Color) -> isize;
-    fn queen_infaltration(&mut self, clr: Color) -> isize;
-
-    // NOTE: 6. MOBILITY Evaluation
-    fn mobility_eval(&mut self, clr: Color);
-    fn mobility_bonus(&mut self, piece: Piece, sq: usize) -> (isize, isize);
-    fn mobility_area(&mut self, clr: Color) -> u64;
-    fn mobility_piece(&mut self, sq: usize, piece: Piece, clr: Color) -> u64;
-
-    // NOTE: 7. THREATS Evaluation TODO: FIXME:
-    fn threats_eval(&mut self, clr: Color);
-    fn safe_pawn(&mut self, clr: Color) -> u64;
-    fn threat_safe_pawn(&mut self, clr: Color) -> u64;
-    fn weak_enemy(&mut self, clr: Color) -> u64;
-    fn minor_threat(&mut self, clr: Color);
-    fn rook_threat(&mut self, clr: Color);
-    fn hanging(&mut self, clr: Color) -> u64;
-    fn king_threat(&mut self, clr: Color) -> u64;
-    fn slider_on_queen(&mut self, clr: Color) -> isize;
-    fn knight_on_queen(&mut self, clr: Color) -> isize;
-    fn restricted(&mut self, clr: Color) -> u64;
-    fn weak_queen_protection(&mut self, clr: Color) -> u64;
-    fn pawn_push_threat(&mut self, clr: Color) -> u64;
-
-    // NOTE: 8. PASSED PAWNS Evaluation
-    fn passed_pawn(&mut self, clr: Color);
-    fn passed_leverable(&mut self, sq: usize, clr: Color) -> bool;
-    fn passed_file(&mut self, sq: usize) -> isize;
-    fn passed_blocked(&mut self, sq: usize, clr: Color) -> isize;
-    fn king_proximity(&mut self, sq: usize, clr: Color) -> isize;
-    fn candidate_passed(&mut self, sq: usize, clr: Color) -> bool;
-
-    // NOTE: 9. SPACE Evaluation FIXME:
-    fn space(&mut self, color: Color);
-    fn space_area(&mut self, color: Color) -> usize;
-
-    // NOTE: 10. KING Evaluation TODO: FIXME:
-    fn king_eval(&mut self, clr: Color);
-    fn king_danger(&mut self, clr: Color) -> isize;
-    fn flank_defense(&mut self, clr: Color) -> isize;
-    fn flank_attack(&mut self, clr: Color) -> isize;
-    fn king_blockers(&mut self, clr: Color);
-    fn endgame_shelter(&mut self, clr: Color) -> isize;
-    fn knight_defender(&mut self, clr: Color) -> u64;
-    fn unsafe_checks(&mut self, clr: Color) -> u64;
-    fn weak_squares(&mut self, clr: Color) -> u64;
-    fn weak_bonus(&mut self, clr: Color) -> u64;
-    fn king_attacks(&mut self, clr: Color) -> isize;
-    fn king_attackers_weight(&mut self, clr: Color) -> isize;
-    fn king_attackers_count(&mut self, clr: Color) -> isize;
-    fn safe_check(&mut self, clr: Color, piece: Piece) -> u64;
-    fn check(&mut self, clr: Color);
-    // fn king_pawn_distance(&mut self, clr: Color);
-    fn shelter(&mut self, clr: Color) -> (isize, isize, isize);
-    // fn shelter_storm(&mut self, clr: Color);
-    // fn shelter_strength(&mut self, clr: Color);
-    fn storm_square(&mut self, sq: usize, clr: Color) -> (isize, isize);
-    fn strength_square(&mut self, sq: usize, clr: Color) -> isize;
-    fn pawnless_flank(&mut self, sq: usize, clr: Color) -> bool;
-
-    // NOTE: 11. TEMPO Evaluation
-    fn tempo(&mut self, color: Color);
-
-    fn get_mask(&mut self, piece: Piece, sq: usize) -> u64;
-    fn x_ray_mask(&mut self, piece: Piece, sq: usize) -> u64;
-
-    fn opp_color_bishops(&mut self, clr: Color) -> bool;
-    fn king_dist(&mut self, clr: Color, sq: usize) -> usize;
-    fn king_ring(&mut self, clr: Color) -> u64;
 }
 
 impl EvaluationTrait for Board {
-    // ************************************************
-    //                     INIT                       *
-    // ************************************************
-
-    #[inline(always)]
-    fn init(&mut self) {
-        self.eval.reset();
-        self.determine_phase();
-
-        self.pawn_init();
-        self.piece_init();
-        self.king_init();
-    }
-
-    #[inline(always)]
-    fn pawn_init(&mut self) {
-        for &clr in &COLORS {
-            let (own, enemy) = self.both_occ_bb(clr);
-            let piece = PAWN + clr;
-            let mut bb = self.bb(piece);
-
-            while let Some(sq) = bb.next() {
-                self.eval.pawn_behind_masks[clr.idx()] |=
-                    PAWN_3_BEHIND_MASKS[clr.idx()][sq] & CLR_CENTER[clr.idx()];
-
-                if !self.backward_pawn(sq, clr)
-                    && !self.blocked_pawn(sq, clr, self.pawn_bb(clr.opp()))
-                {
-                    self.eval.pawn_att_span[clr.idx()] |=
-                        FORWARD_SPANS_LR[clr.idx()][sq] | get_pawn_att_mask(sq, own, enemy, clr);
-                }
-
-                self.eval.king_pawn_dx[clr.idx()] =
-                    self.eval.king_pawn_dx[clr.idx()].min(self.king_dist(clr, sq));
-
-                self.eval.open_file[clr.idx()] &= !FILE_BITBOARD[get_file(sq)]; // OPEN-FILE
-            }
-        }
-
-        for &clr in &COLORS {
-            self.eval.outpost[clr.idx()] = !self.eval.pawn_att_span[clr.opp().idx()]
-                & OUTPOST_RANKS[clr.idx()]
-                & (get_all_pawn_left_att_mask(self.pawn_bb(clr), clr)
-                    | get_all_pawn_right_att_mask(self.pawn_bb(clr), clr));
-        }
-    }
-
-    #[inline(always)]
-    fn piece_init(&mut self) {
-        for &clr in &COLORS {
-            let (own, enemy) = self.both_occ_bb(clr);
-            let area = self.mobility_area(clr);
-            let king_sq = self.king_sq(clr.opp());
-            let king_ring = self.king_ring(clr.opp());
-
-            let opp_king_mask = get_king_mask(king_sq, 0, 0, clr.opp());
-
-            for &pce in &PIECES {
-                let piece = pce + clr;
-                let mut bb = self.bb(piece);
-                let mut attckers_count = 0;
-
-                while let Some(sq) = bb.next() {
-                    let piece_mask = self.x_ray_mask(piece, sq);
-
-                    self.eval.attacked_by_2[clr.idx()] |=
-                        self.eval.attack_map[clr.idx()] & (piece_mask & !own);
-
-                    self.eval.defended_by_2[clr.idx()] |=
-                        self.eval.defend_map[clr.idx()] & (piece_mask & own);
-
-                    self.eval.attack_map[clr.idx()] |= piece_mask & !own;
-                    self.eval.defend_map[clr.idx()] |= piece_mask & own;
-
-                    self.eval.attacked_by[piece.idx()] |= piece_mask;
-
-                    match piece.kind() {
-                        PAWN => {
-                            if piece_mask & KING_RING[king_sq] != 0 {
-                                attckers_count += 1;
-                                self.eval.king_att_count[clr.idx()] +=
-                                    (piece_mask & KING_RING[king_sq]).count();
-                            }
-                        }
-                        KING => {}
-                        _ => {
-                            let safe_squares = (self.mobility_piece(sq, piece, clr) & area).count();
-                            self.eval.mobility[clr.idx()] +=
-                                self.mobility_bonus(piece, safe_squares).0;
-
-                            if piece_mask & king_ring != 0 {
-                                attckers_count += 1;
-                                self.eval.king_att_count[clr.idx()] += 1;
-                                self.eval.king_att_count_pieces[clr.idx()].set_bit(sq);
-                            }
-
-                            self.eval.king_att[clr.idx()] += (piece_mask & opp_king_mask).count();
-
-                            self.eval.x_ray[piece.idx()] |= self.mobility_piece(sq, piece, clr);
-                        }
-                    }
-
-                    if piece.is_queen() {
-                        self.eval.queen_diagonal[clr.idx()] |= get_bishop_mask(sq, own, enemy, clr);
-                    }
-                }
-
-                self.eval.king_att_weight[clr.idx()] +=
-                    KING_ATT_WEIGHT[piece.arr_idx()] * attckers_count as isize;
-            }
-        }
-    }
-
-    #[inline(always)]
-    fn king_init(&mut self) {
-        for clr in COLORS {
-            self.eval.king_ring[clr.idx()] = self.king_ring(clr);
-            self.check(clr);
-            self.eval.king_shelter[clr.idx()] = self.shelter(clr);
-        }
-    }
-
-    #[inline(always)]
-    fn insufficient_material(&self) -> bool {
-        // The Color is not relevant here, and that is why i use self.color()
-        let (own, enemy) = self.both_occ_bb(self.color());
-        if (own | enemy).count_ones() < 4 {
-            let kings = self.bb(WHITE_KING) | self.bb(BLACK_KING);
-            let knights = self.bb(WHITE_KNIGHT) | self.bb(BLACK_KNIGHT);
-            let bishops = self.bb(WHITE_BISHOP) | self.bb(BLACK_BISHOP);
-            if (kings | knights | bishops) == (own | enemy) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    #[inline(always)]
-    fn determine_phase(&mut self) {
-        let mut npm = self.non_pawn_material_eval(WHITE) + self.non_pawn_material_eval(BLACK);
-        // println!("{:?}", self.non_pawn_material_eval(WHITE));
-        // println!("{:?}", npm);
-
-        npm = EG_LIMIT.max(npm.min(MG_LIMIT));
-        // println!("{:?}", npm);
-
-        let phase = ((npm - EG_LIMIT) * 128) / (MG_LIMIT - EG_LIMIT);
-        // println!("{:?}", (npm, phase));
-
-        self.eval.phase = (phase, 128 - phase);
-        // println!("{:?}", (phase, 128 - phase));
-    }
-
-    #[inline(always)]
-    fn tapered(&mut self, value: (isize, isize)) -> isize {
-        (self.eval.phase.0 * value.0 + self.eval.phase.1 * value.1) / 128
-    }
-
-    #[inline(always)]
-    fn calculate_score(&mut self) -> isize {
-        let mg = self.eval.score[WHITE.idx()].0 - self.eval.score[BLACK.idx()].0;
-        let eg = self.eval.score[WHITE.idx()].1 - self.eval.score[BLACK.idx()].1;
-
-        return self.tapered((mg, eg));
-    }
-
-    #[inline(always)]
-    fn front_sq(&mut self, sq: usize, clr: Color) -> usize {
-        (sq as isize + 8 * clr.sign()) as usize
-    }
-
-    #[inline(always)]
-    fn back_sq(&mut self, sq: usize, clr: Color) -> usize {
-        (sq as isize - 8 * clr.sign()) as usize
-    }
-
-    #[inline(always)]
-    fn pin_att(&mut self, from: usize, to: usize, piece: Piece) -> u64 {
-        let (from_file, from_rank) = (get_file(from), get_rank(from));
-        let (to_file, to_rank) = (get_file(to), get_rank(to));
-
-        match piece {
-            p if p.is_queen() => BETWEEN_BB[from][to],
-            p if p.is_bishop() => {
-                let diagonal_move = from_file != to_file && from_rank != to_rank;
-                BETWEEN_BB[from][to] * diagonal_move as u64
-            }
-            p if p.is_rook() => {
-                let straight_move = from_file == to_file || from_rank == to_rank;
-                BETWEEN_BB[from][to] * straight_move as u64
-            }
-            _ => 0,
-        }
-    }
-
-    #[inline(always)]
-    fn sum(
-        &mut self,
-        color: Color,
-        square: Option<usize>,
-        piece: Option<Piece>,
-        value: (isize, isize),
-    ) {
-        // self.trace(color, square, piece, value);
-
-        self.eval.score[color.idx()].0 += value.0;
-        self.eval.score[color.idx()].1 += value.1;
-    }
-
-    #[inline(always)]
-    fn trace(
-        &mut self,
-        color: Color,
-        square: Option<usize>,
-        piece: Option<Piece>,
-        value: (isize, isize),
-    ) {
-        self.eval.vec_test.push(format!(
-            "Piece:{:?} ,Color: {:?}, Square: {:?}, Value Mg: {:?}, Value Eg: {:?}",
-            piece, color, square, value.0, value.1
-        ));
-
-        if let Some(sq) = square {
-            self.eval.mg_test[color.idx()][sq] += value.0;
-            self.eval.eg_test[color.idx()][sq] += value.1;
-        }
-    }
-
-    #[inline(always)]
-    fn reset_trace(&mut self) {
-        self.eval.vec_test.clear();
-        self.eval.mg_test = [[0; 64]; 2];
-        self.eval.eg_test = [[0; 64]; 2];
-    }
-
-    #[inline(always)]
-    fn print_trace_board(&mut self, name: &str) {
-        let mg_test = self.eval.mg_test.map(|row| row.map(|num| num.to_string()));
-        let eg_test = self.eval.eg_test.map(|row| row.map(|num| num.to_string()));
-
-        println!("--------------Print Evaluation Board for: {:?}-------------", name);
-        println!("{:?}", "");
-        println!("******* Color White, Phase: Middle Game *******");
-        print_eval(&mg_test[WHITE.idx()]);
-        println!("******* Color Black, Phase: Middle Game *******");
-        print_eval(&mg_test[BLACK.idx()]);
-        println!("******* Color White, Phase: End Game *******");
-        print_eval(&eg_test[WHITE.idx()]);
-        println!("******* Color Black, Phase: End Game *******");
-        print_eval(&eg_test[BLACK.idx()]);
-    }
-
-    #[inline(always)]
-    fn print_trace_log(&mut self, name: &str) {
-        println!("--------------Print Evaluation Log for: {:?}-------------", name);
-        for log in &self.eval.vec_test {
-            println!("{:?}", log);
-        }
-    }
-
-    #[inline(always)]
-    fn print_trace_score(&mut self, name: &str) {
-        println!("-------------- Print Evaluation Score for: {:?} -------------", name);
-        println!("-> Color White, Phase: Mg, Score: {:?} ", self.eval.score[WHITE.idx()].0);
-        println!("-> Color Black, Phase: Mg, Score: {:?} ", self.eval.score[BLACK.idx()].0);
-        println!("-> Color White, Phase: Eg, Score: {:?} ", self.eval.score[WHITE.idx()].1);
-        println!("-> Color Black, Phase: Eg, Score: {:?} ", self.eval.score[BLACK.idx()].1);
-    }
-
-    // ************************************************
-    //                    ATTACKS                     *
-    // ************************************************
-
-    #[inline(always)]
-    fn x_ray_mask(&mut self, piece: Piece, sq: usize) -> u64 {
-        let clr = piece.color();
-        let (mut own, mut enemy) = self.both_occ_bb(clr);
-        match piece.kind() {
-            PAWN => get_pawn_att_mask(sq, own, enemy, clr),
-            KNIGHT => get_knight_mask(sq, own, enemy, clr),
-            BISHOP => {
-                own &= !(self.queen_bb(clr));
-                enemy &= !self.queen_bb(clr.opp());
-                get_bishop_mask(sq, own, enemy, clr)
-            }
-            ROOK => {
-                own &= !(self.queen_bb(clr) | self.rook_bb(clr));
-                enemy &= !self.queen_bb(clr.opp());
-                get_rook_mask(sq, own, enemy, clr)
-            }
-            QUEEN => get_queen_mask(sq, own, enemy, clr),
-            KING => get_king_mask(sq, own, enemy, clr),
-            _ => panic!("Invalid Peace Type"),
-        }
-    }
-
-    #[inline(always)]
-    fn get_mask(&mut self, piece: Piece, sq: usize) -> u64 {
-        let (own, enemy) = self.both_occ_bb(piece.color());
-        match piece.kind() {
-            PAWN => get_pawn_att_mask(sq, own, enemy, piece.color()),
-            KNIGHT => get_knight_mask(sq, own, enemy, piece.color()),
-            BISHOP => get_bishop_mask(sq, own, enemy, piece.color()),
-            ROOK => get_rook_mask(sq, own, enemy, piece.color()),
-            QUEEN => get_queen_mask(sq, own, enemy, piece.color()),
-            KING => get_king_mask(sq, own, enemy, piece.color()),
-            _ => panic!("Invalid Peace Type"),
-        }
-    }
-
     // ************************************************
     //                MAIN EVALUATION                 *
     // ************************************************
@@ -678,1800 +214,15 @@ impl EvaluationTrait for Board {
 
         return self.calculate_score() * self.color().sign();
     }
-
-    // ************************************************
-    //           1. MATERIAL EVALUATION               *
-    // ************************************************
-
-    #[inline(always)]
-    fn material_eval(&mut self, clr: Color) {
-        for &pce in &PIECES {
-            let piece = pce + clr;
-            let count = self.bb(piece).count() as isize;
-            let (mg_sum, eg_sum) = self.piece_material(piece);
-            self.sum(clr, None, Some(piece), (mg_sum * count, eg_sum * count));
-        }
-    }
-
-    #[inline(always)]
-    fn non_pawn_material_eval(&mut self, clr: Color) -> isize {
-        let mut score = 0;
-        for &pce in &PIECES_WITHOUT_PAWN {
-            let piece = pce + clr;
-            let count = self.bb(piece).count() as isize;
-            score += self.piece_material(piece).0 * count;
-        }
-        score
-    }
-
-    #[inline(always)]
-    fn piece_material(&mut self, piece: Piece) -> (isize, isize) {
-        PIECE_MATERIAL[piece.arr_idx()]
-    }
-
-    // ************************************************
-    //             2. PSQT EVALUATION                 *
-    // ************************************************
-
-    #[inline(always)]
-    fn psqt_eval(&mut self, clr: Color) {
-        for &pce in &PIECES {
-            let piece = pce + clr;
-            let mut bb = self.bb(piece);
-            while let Some(sq) = bb.next() {
-                let bonus = self.piece_psqt(piece, sq);
-                self.sum(piece.color(), Some(sq), Some(piece + clr), bonus);
-            }
-        }
-    }
-
-    #[inline(always)]
-    fn piece_psqt(&mut self, piece: Piece, sq: usize) -> (isize, isize) {
-        let fixed_sq = CLR_SQ[piece.color().idx()][sq];
-        PSQT[piece.arr_idx()][fixed_sq]
-    }
-
-    // ************************************************
-    //            3. IMBALANCE EVALUATION             *
-    // ************************************************
-
-    fn imbalance(&mut self, clr: Color) {
-        let ours: [isize; 6] = [
-            0,
-            self.pawn_bb(clr).count() as isize,
-            self.knight_bb(clr).count() as isize,
-            self.bishop_bb(clr).count() as isize,
-            self.rook_bb(clr).count() as isize,
-            self.queen_bb(clr).count() as isize,
-        ];
-        let theirs: [isize; 6] = [
-            0,
-            self.pawn_bb(clr.opp()).count() as isize,
-            self.knight_bb(clr.opp()).count() as isize,
-            self.bishop_bb(clr.opp()).count() as isize,
-            self.rook_bb(clr.opp()).count() as isize,
-            self.queen_bb(clr.opp()).count() as isize,
-        ];
-        let mut bonus = 0;
-
-        let has_our_bishop_pair = has_bishop_pair(self.bishop_bb(clr)) as isize;
-        let has_their_bishop_pair = has_bishop_pair(self.bishop_bb(clr.opp())) as isize;
-
-        for pt1 in 1..6 {
-            if ours[pt1] == 0 {
-                continue;
-            }
-
-            bonus += (QUADRATIC_OURS[pt1][0] * has_our_bishop_pair
-                + QUADRATIC_OURS[pt1][1] * ours[1]
-                + QUADRATIC_OURS[pt1][2] * ours[2]
-                + QUADRATIC_OURS[pt1][3] * ours[3]
-                + QUADRATIC_OURS[pt1][4] * ours[4]
-                + QUADRATIC_OURS[pt1][5] * ours[5]
-                + QUADRATIC_THEIRS[pt1][0] * has_their_bishop_pair
-                + QUADRATIC_THEIRS[pt1][1] * theirs[1]
-                + QUADRATIC_THEIRS[pt1][2] * theirs[2]
-                + QUADRATIC_THEIRS[pt1][3] * theirs[3]
-                + QUADRATIC_THEIRS[pt1][4] * theirs[4]
-                + QUADRATIC_THEIRS[pt1][5] * theirs[5])
-                * ours[pt1];
-        }
-
-        bonus += 1438 * has_our_bishop_pair;
-        bonus /= 16;
-        self.sum(clr, None, None, (bonus, bonus));
-    }
-
-    #[inline(always)]
-    fn imb_piece_count(&mut self, num: usize, clr: Color) -> isize {
-        match num {
-            0 => 0, //self.king_bb(clr).count() as isize,
-            1 => self.pawn_bb(clr).count() as isize,
-            2 => self.knight_bb(clr).count() as isize,
-            3 => self.bishop_bb(clr).count() as isize,
-            4 => self.rook_bb(clr).count() as isize,
-            5 => self.queen_bb(clr).count() as isize,
-            _ => panic!("Sth is not right"),
-        }
-    }
-
-    // ************************************************
-    //              4. PAWNS EVALUATION               *
-    // ************************************************
-
-    #[inline(always)] // 4. Pawns Eval
-    fn pawns_eval(&mut self, clr: Color) {
-        let mut bb = self.pawn_bb(clr);
-        while let Some(sq) = bb.next() {
-            self.single_pawn_eval(sq, clr);
-        }
-    }
-
-    #[inline(always)]
-    fn single_pawn_eval(&mut self, sq: usize, clr: Color) {
-        // println!("Sq: {:?}", sq);
-        // println!("Double Isolated: {:?}", self.doubled_isolated_pawn(sq, clr));
-        // println!("Isolated: {:?}", self.isolated_pawn(sq, clr));
-        // println!("Backward Pawn: {:?}", self.backward_pawn(sq, clr));
-        // println!("Doubled Pawn: {:?}", self.doubled_pawn(sq, clr));
-        // println!("Connected Pawn: {:?}", self.connected_pawn(sq, clr));
-        // println!("Weak Unopposed Pawn: {:?}", self.weak_unopposed_pawn(sq, clr));
-        // println!("Weak Lever: {:?}", self.weak_lever(sq, clr));
-        // println!("Blocked pawn 5th 6th: {:?}", self.blocked_pawn_5th_6th_rank(sq, clr));
-        // println!("Connected: {:?}", self.blocked_pawn_5th_6th_rank(sq, clr));
-
-        if self.doubled_isolated_pawn(sq, clr) {
-            self.sum(clr, Some(sq), None, (-11, -56));
-        } else if self.isolated_pawn(sq, clr) {
-            self.sum(clr, Some(sq), None, (-5, -15));
-        } else if self.backward_pawn(sq, clr) {
-            self.sum(clr, Some(sq), None, (-9, -24));
-        }
-
-        // FIXME: Not correct (Needs to check how many doubled are on the same file)
-        if self.doubled_pawn(sq, clr) {
-            self.sum(clr, Some(sq), None, (-11, -56));
-        }
-
-        if self.connected_pawn(sq, clr) {
-            let calc_bonus = self.connected_bonus(sq, clr);
-            let bonus =
-                (calc_bonus, calc_bonus * (CLR_RANK[clr.idx()][get_rank(sq)] as isize - 2) / 4);
-            self.sum(clr, Some(sq), None, bonus);
-            // FIXME: Check if it is ok to be this a minus sth
-            // println!("Connected Bonus: {:?}", bonus);
-        }
-
-        if self.weak_unopposed_pawn(sq, clr) {
-            self.sum(clr, Some(sq), None, (-13, -27));
-        }
-
-        if self.weak_lever(sq, clr) {
-            self.sum(clr, Some(sq), None, (0, -56));
-        }
-
-        if self.blocked_pawn_5th_6th_rank(sq, clr) == 1 {
-            self.sum(clr, Some(sq), None, (-11, -4));
-        } else if self.blocked_pawn_5th_6th_rank(sq, clr) == 2 {
-            self.sum(clr, Some(sq), None, (-3, 4));
-        }
-    }
-
-    #[inline(always)]
-    fn isolated_pawn(&mut self, sq: usize, clr: Color) -> bool {
-        ISOLATED_PAWN_LOOKUP[sq] & self.pawn_bb(clr) == 0
-    }
-
-    #[inline(always)]
-    fn opposed_pawn(&mut self, sq: usize, clr: Color) -> bool {
-        PAWN_FORWARD_SPANS[clr.idx()][sq] & self.pawn_bb(clr.opp()) != 0
-    }
-
-    #[inline(always)]
-    fn phalanx_pawn(&mut self, sq: usize, clr: Color) -> bool {
-        PAWN_ATTACK_LOOKUP[clr.opp().idx()][(sq as isize + 8 * clr.sign()) as usize]
-            & self.pawn_bb(clr)
-            != 0
-    }
-
-    #[inline(always)]
-    fn supported_pawn(&mut self, sq: usize, clr: Color) -> isize {
-        (PAWN_ATTACK_LOOKUP[clr.opp().idx()][sq] & self.pawn_bb(clr)).count() as isize
-    }
-
-    #[inline(always)]
-    fn backward_pawn(&mut self, sq: usize, clr: Color) -> bool {
-        let front_sq = self.front_sq(sq, clr);
-        ((FORWARD_SPANS_LR[clr.opp().idx()][sq]
-            | PAWN_ATTACK_LOOKUP[clr.opp().idx()][(sq as isize + 8 * clr.sign()) as usize])
-            & self.pawn_bb(clr)
-            == 0)
-            && (self.blocked_pawn(sq, clr, self.pawn_bb(clr.opp()))
-                || self.eval.attacked_by[(PAWN + clr.opp()).idx()].is_set(front_sq))
-    }
-
-    #[inline(always)]
-    fn doubled_pawn(&mut self, sq: usize, clr: Color) -> bool {
-        self.pawn_bb(clr).is_set(self.back_sq(sq, clr)) && self.supported_pawn(sq, clr) == 0
-        // PAWN_FORWARD_SPANS[clr.opp().idx()][sq] & self.pawn_bb(clr) != 0
-    }
-
-    #[inline(always)]
-    fn connected_pawn(&mut self, sq: usize, clr: Color) -> bool {
-        self.supported_pawn(sq, clr) > 0 || self.phalanx_pawn(sq, clr)
-    }
-
-    #[inline(always)]
-    fn connected_bonus(&mut self, sq: usize, clr: Color) -> isize {
-        if !self.connected_pawn(sq, clr) {
-            return 0;
-        }
-
-        let r = CLR_RANK[clr.idx()][get_rank(sq)];
-        if r < 1 || r > 6 {
-            return 0;
-        }
-
-        let seed = [0, 7, 8, 12, 29, 48, 86, 0];
-        let op = self.opposed_pawn(sq, clr);
-        let ph = self.phalanx_pawn(sq, clr);
-        let su = self.supported_pawn(sq, clr);
-        let bl = self.blocked_pawn(sq, clr, self.pawn_bb(clr.opp()));
-
-        // println!("Sq: {:?}", sq);
-        // println!("Opposed Pawn: {:?}", self.opposed_pawn(sq, clr));
-        // println!("Phalanx: {:?}", self.phalanx_pawn(sq, clr));
-        // println!("Supported: {:?}", self.supported_pawn(sq, clr));
-        // println!("Blocked Pawn: {:?}", self.doubled_pawn(sq, clr));
-        // println!("Bonus: {:?}", seed[r] * (2 + ph as isize - op as isize) + 21 * su as isize);
-
-        return seed[r] * (2 + ph as isize - op as isize) + 21 * su as isize;
-    }
-
-    #[inline(always)]
-    fn weak_unopposed_pawn(&mut self, sq: usize, clr: Color) -> bool {
-        !self.opposed_pawn(sq, clr) && (self.isolated_pawn(sq, clr) || self.backward_pawn(sq, clr))
-    }
-
-    #[inline(always)]
-    fn weak_lever(&mut self, sq: usize, clr: Color) -> bool {
-        self.supported_pawn(sq, clr) == 0
-            && (get_pawn_att_mask(sq, 0, 0, clr) & self.pawn_bb(clr.opp())).count() == 2
-    }
-
-    #[inline(always)]
-    fn blocked_pawn(&mut self, sq: usize, clr: Color, bb: u64) -> bool {
-        get_all_pawn_forward_mask(Bitboard::init(sq), clr) & bb != 0
-    }
-
-    #[inline(always)] // Blocked only on the 5th and 6 rank
-    fn blocked_pawn_5th_6th_rank(&mut self, sq: usize, clr: Color) -> isize {
-        if BLOCKED_RANKS[clr.idx()].is_set(sq)
-            && self.blocked_pawn(sq, clr, self.pawn_bb(clr.opp()))
-        {
-            return CLR_RANK[clr.idx()][get_rank(sq)].abs_diff(3) as isize;
-        }
-        return 0;
-    }
-
-    #[inline(always)]
-    fn doubled_isolated_pawn(&mut self, sq: usize, clr: Color) -> bool {
-        PAWN_FORWARD_SPANS[clr.opp().idx()][sq] & self.pawn_bb(clr) != 0
-            && self.opposed_pawn(sq, clr)
-            && self.isolated_pawn(sq, clr)
-            && self.isolated_pawn(sq, clr.opp())
-
-        // self.doubled_pawn(sq, clr)
-        //     && self.blocked_pawn(sq, clr, self.pawn_bb(clr.opp()))
-        //     && self.isolated_pawn(sq, clr)
-        //     && self.isolated_pawn((sq as isize + 8 * clr.sign()) as usize, clr)
-    }
-
-    // ************************************************
-    //              5. PIECE EVALUATION               *
-    // ************************************************
-
-    #[inline(always)]
-    fn piece_eval(&mut self, clr: Color) {
-        let bonus = self.minor_behind_pawn(clr);
-        self.sum(clr, None, None, (18 * bonus, 3 * bonus));
-
-        let bonus = self.bishop_pawns(clr);
-        self.sum(clr, None, None, (-3 * bonus, -5 * bonus));
-
-        let bonus = self.bishop_xray_pawns(clr);
-        self.sum(clr, None, None, (-4 * bonus, -5 * bonus));
-
-        self.rook_on_queen_file(clr);
-
-        let bonus = self.rook_on_king_ring(clr);
-        self.sum(clr, None, None, (16 * bonus, 0));
-
-        let bonus = self.bishop_on_king_ring(clr);
-        self.sum(clr, None, None, (24 * bonus, 0));
-
-        self.trapped_rook(clr);
-        // FIXME, If Castle is not awailable add 2 * 55 / 2 * 13
-        // self.sum(clr, None, None, (-55 * bonus, -13 * bonus));
-
-        let bonus = self.weak_queen(clr);
-        self.sum(clr, None, None, (-56 * bonus, -15 * bonus));
-
-        let bonus = self.queen_infaltration(clr);
-        self.sum(clr, None, None, (-2 * bonus, 14 * bonus));
-
-        self.king_protector(clr);
-
-        self.outpost_total(clr);
-
-        let bonus = (self.rook_bb(clr)
-            & (self.eval.open_file[clr.idx()] & self.eval.open_file[clr.opp().idx()]))
-        .count() as isize;
-        self.sum(clr, None, None, (48 * bonus, 29 * bonus));
-
-        let bonus = (self.rook_bb(clr) & self.eval.open_file[clr.idx()]).count() as isize;
-        self.sum(clr, None, None, (19 * bonus, 7 * bonus));
-    }
-
-    #[inline(always)]
-    fn outpost(&mut self, clr: Color) -> u64 {
-        (self.knight_bb(clr) | self.bishop_bb(clr)) & self.eval.outpost[clr.idx()]
-    }
-
-    #[inline(always)]
-    fn reachable_outpost(&mut self, clr: Color) -> u64 {
-        let att = self.eval.attacked_by[(KNIGHT + clr).idx()]
-            | self.eval.attacked_by[(BISHOP + clr).idx()];
-        self.eval.outpost[clr.idx()] & !self.occ_bb(clr) & att
-        // let reachable_bb = self.eval.outpost[clr.idx()] & !self.occ_bb(clr) & att;
-        // (reachable_bb.count() * 2) as isize
-    }
-
-    #[inline(always)]
-    fn minor_behind_pawn(&mut self, clr: Color) -> isize {
-        let all_pawns = self.pawn_bb(clr) | self.pawn_bb(clr.opp());
-        ((self.knight_bb(clr) | self.bishop_bb(clr))
-            & get_all_pawn_forward_mask(all_pawns, clr.opp()))
-        .count() as isize
-    }
-
-    #[inline(always)]
-    fn bishop_pawns(&mut self, clr: Color) -> isize {
-        let mut score = 0;
-        let mut blocked = get_all_pawn_forward_mask(self.pawn_bb(clr), clr)
-            & (self.occ_bb(clr) | self.occ_bb(clr.opp()))
-            & (FILE_BITBOARD[2] | FILE_BITBOARD[3] | FILE_BITBOARD[4] | FILE_BITBOARD[5]);
-        blocked = get_all_pawn_forward_mask(blocked, clr.opp());
-        // print_bitboard(blocked, None);
-
-        for squares in [WHITE_SQUARES, BLACK_SQUARES] {
-            let bishops_on_sq = self.bishop_bb(clr) & squares;
-            // print_bitboard(bishops_on_sq, None);
-
-            let pawns_on_sq = self.pawn_bb(clr) & squares;
-            // print_bitboard(pawns_on_sq, None);
-
-            // let blocked_on_sq = blocked & squares;
-            // print_bitboard(blocked_on_sq, None);
-
-            let att_bishops =
-                if self.eval.attacked_by[(PAWN + clr).idx()] & bishops_on_sq > 0 { 0 } else { 1 };
-            // print_bitboard(att_bishops, None);
-
-            score += pawns_on_sq.count()
-                * (blocked.count() * bishops_on_sq.count() + att_bishops * bishops_on_sq.count());
-
-            // println!("Temp Score: {:?}", score);
-        }
-
-        // println!("Final Score: {:?}", score);
-        return score as isize;
-    }
-
-    #[inline(always)]
-    fn trapped_rook(&mut self, clr: Color) {
-        let mut bb = self.rook_bb(clr) & !self.eval.open_file[clr.idx()];
-        let king_file = get_file(self.king_sq(clr));
-        while let Some(sq) = bb.next() {
-            if (self.x_ray_mask(ROOK + clr, sq).count() <= 3)
-                && ((king_file < 4) == (get_file(sq) < king_file))
-            {
-                let mut castling = 2;
-                if self.state.castling.long(clr) != 0 || self.state.castling.short(clr) != 0 {
-                    castling = 1;
-                }
-                self.sum(clr, Some(sq), Some(ROOK + clr), (-55 * castling, -13 * castling));
-            }
-        }
-    }
-
-    #[inline(always)]
-    fn weak_queen(&mut self, clr: Color) -> isize {
-        let mut bb = self.queen_bb(clr);
-        let mut count = 0;
-        while let Some(sq_to) = bb.next() {
-            let mut bb = (get_rook_mask(sq_to, 0, 0, clr) & self.rook_bb(clr.opp()))
-                | (get_bishop_mask(sq_to, 0, 0, clr) & self.bishop_bb(clr.opp()));
-
-            while let Some(sq_from) = bb.next() {
-                if BETWEEN_BB[sq_from][sq_to].count() == 3 {
-                    count += 1;
-                    break;
-                }
-            }
-        }
-
-        return count;
-    }
-
-    #[inline(always)]
-    fn king_protector(&mut self, clr: Color) {
-        let mut bb = self.knight_bb(clr);
-        while let Some(sq) = bb.next() {
-            let dx = self.king_dist(clr, sq) as isize;
-            self.sum(clr, Some(sq), Some(KNIGHT), (-8 * dx, -9 * dx));
-        }
-
-        let mut bb = self.bishop_bb(clr);
-        while let Some(sq) = bb.next() {
-            let dx = self.king_dist(clr, sq) as isize;
-            self.sum(clr, Some(sq), Some(BISHOP), (-6 * dx, -9 * dx));
-        }
-    }
-
-    #[inline(always)]
-    fn outpost_total(&mut self, clr: Color) {
-        let mut bb = self.knight_bb(clr);
-        while let Some(sq) = bb.next() {
-            let reachable_bb = self.eval.outpost[clr.idx()]
-                & self.x_ray_mask(KNIGHT + clr, sq)
-                & !self.occ_bb(clr);
-            if !self.eval.outpost[clr.idx()].is_set(sq) && reachable_bb > 0 {
-                self.sum(clr, Some(sq), Some(KNIGHT + clr), (31, 22));
-                break;
-            }
-        }
-
-        let bonus = (self.knight_bb(clr) & self.eval.outpost[clr.idx()]).count() as isize;
-        self.sum(clr, None, Some(KNIGHT + clr), (bonus * 56, bonus * 36));
-
-        let bonus = (self.bishop_bb(clr) & self.eval.outpost[clr.idx()]).count() as isize;
-        self.sum(clr, None, Some(BISHOP + clr), (bonus * 30, bonus * 23));
-        // NOTE: FIXME: NOT FULL EVAL BUT AN OK ONE
-        // Only the +2 is missing
-    }
-
-    #[inline(always)]
-    fn rook_on_queen_file(&mut self, clr: Color) {
-        let mut bb = self.rook_bb(clr);
-        let all_queens = self.queen_bb(clr) | self.queen_bb(clr.opp());
-        while let Some(sq) = bb.next() {
-            if all_queens & FILE_BITBOARD[get_file(sq)] != 0 {
-                self.sum(clr, Some(sq), Some(ROOK), (6, 11));
-            }
-        }
-    }
-
-    #[inline(always)]
-    fn bishop_xray_pawns(&mut self, clr: Color) -> isize {
-        let mut count = 0;
-        let mut bb = self.bishop_bb(clr);
-        while let Some(sq) = bb.next() {
-            count += (get_bishop_mask(sq, 0, 0, clr) & self.pawn_bb(clr.opp())).count();
-        }
-
-        return count as isize;
-    }
-
-    #[inline(always)]
-    fn rook_on_king_ring(&mut self, clr: Color) -> isize {
-        let mut count = 0;
-        let mut bb = self.rook_bb(clr) & !self.eval.king_att_count_pieces[clr.idx()];
-        while let Some(sq) = bb.next() {
-            if self.eval.king_ring[clr.opp().idx()] & FILE_BITBOARD[get_file(sq)] != 0 {
-                count += 1;
-            }
-        }
-        count
-    }
-
-    #[inline(always)]
-    fn bishop_on_king_ring(&mut self, clr: Color) -> isize {
-        let mut count = 0;
-        let mut bb = self.bishop_bb(clr) & !self.eval.king_att_count_pieces[clr.idx()];
-        while let Some(sq) = bb.next() {
-            if self.eval.king_ring[clr.opp().idx()]
-                & get_bishop_mask(sq, self.pawn_bb(clr), self.pawn_bb(clr.opp()), clr)
-                != 0
-            {
-                count += 1;
-            }
-        }
-        count
-    }
-
-    #[inline(always)]
-    fn queen_infaltration(&mut self, clr: Color) -> isize {
-        let bb = self.queen_bb(clr)
-            & QUEEN_INFILTRATION[clr.idx()]
-            & !(self.eval.attacked_by[(PAWN + clr.opp()).idx()]
-                | self.eval.defended_by[(PAWN + clr.opp()).idx()])
-            & !self.eval.pawn_att_span[clr.opp().idx()];
-        bb.count() as isize
-    }
-
-    // ************************************************
-    //             6. MOBILITY EVALUATION             *
-    // ************************************************
-
-    #[inline(always)]
-    fn mobility_eval(&mut self, clr: Color) {
-        let area = self.mobility_area(clr);
-        for pce in [KNIGHT, BISHOP, ROOK, QUEEN] {
-            let piece = pce + clr;
-            let mut bb = self.bb(piece);
-            while let Some(sq) = bb.next() {
-                let safe_squares = (self.mobility_piece(sq, piece, clr) & area).count();
-                let bonus = self.mobility_bonus(piece, safe_squares);
-                self.sum(clr, Some(sq), Some(piece), bonus);
-                // self.eval.mobility[clr.idx()] += bonus.0;
-            }
-        }
-    }
-
-    #[inline(always)]
-    fn mobility_bonus(&mut self, piece: Piece, safe_sqaures: usize) -> (isize, isize) {
-        match piece.kind() {
-            KNIGHT => KNIGHT_MOBILITY[safe_sqaures],
-            BISHOP => BISHOP_MOBILITY[safe_sqaures],
-            ROOK => ROOK_MOBILITY[safe_sqaures],
-            QUEEN => QUEEN_MOBILITY[safe_sqaures],
-            _ => panic!("There is other peace that was not expected here"),
-        }
-    }
-
-    #[inline(always)]
-    fn mobility_piece(&mut self, sq: usize, piece: Piece, clr: Color) -> u64 {
-        let (mut own, mut enemy) = self.both_occ_bb(clr);
-        match piece.kind() {
-            KNIGHT => get_knight_mask(sq, own, enemy, clr),
-            BISHOP => {
-                own &= !(self.queen_bb(clr));
-                enemy &= !self.queen_bb(clr.opp());
-                get_bishop_mask(sq, own, enemy, clr)
-            }
-            ROOK => {
-                own &= !(self.queen_bb(clr) | self.rook_bb(clr));
-                enemy &= !self.queen_bb(clr.opp());
-                get_rook_mask(sq, own, enemy, clr)
-            }
-            QUEEN => get_queen_mask(sq, own, enemy, clr),
-            _ => panic!("There is other peace that was not expected here"),
-        }
-    }
-
-    #[inline(always)]
-    fn mobility_area(&mut self, clr: Color) -> u64 {
-        let bb = (u64::MAX)
-            & !self.king_bb(clr)
-            & !self.queen_bb(clr)
-            & !self.pawn_bb(clr)
-            & !get_all_pawn_left_att_mask(self.pawn_bb(clr.opp()), clr.opp())
-            & !get_all_pawn_right_att_mask(self.pawn_bb(clr.opp()), clr.opp());
-        bb
-    }
-
-    // ************************************************
-    //             7. PASSED PAWN EVALUATION          *
-    // ************************************************
-
-    // 8 Passed Pawns
-
-    // fn passed_pawn(&mut self, clr: Color) -> isize {
-    //     let mut score = 0;
-    //     let bb = self.pawn_bb(clr);
-    //     for sq in bb.next() {
-    //         if !passed_leverable(pos, square) {
-    //             return 0;
-    //         }
-
-    //         score += king_proximity(pos, square); // FIXME: Only In endgame
-    //         score += self.tapered(PASSED_PAWN_REW[clr.idx()][get_rank(sq)]);
-    //         score += self.passed_block(pos, square);
-    //         score -= self.tapered((11, 8)) * PSQT_FILE[get_file(sq)] as isize;
-    //     }
-    //     score
-    // }
-    // fn passed_block(&mut self, sq: usize, clr: Color) {
-
-    // }
-
-    // ************************************************
-    //              8. THREATS EVALUATION             *
-    // ************************************************
-    #[inline(always)]
-    fn threats_eval(&mut self, clr: Color) {
-        let bonus = self.hanging(clr).count() as isize;
-        self.sum(clr, None, None, (69 * bonus, 36 * bonus));
-
-        if self.king_threat(clr) > 0 {
-            self.sum(clr, None, Some(KING + clr), (24, 89));
-        }
-
-        let bonus = self.pawn_push_threat(clr).count() as isize;
-        self.sum(clr, None, None, (48 * bonus, 39 * bonus));
-
-        let bonus = self.threat_safe_pawn(clr).count() as isize;
-        self.sum(clr, None, None, (173 * bonus, 94 * bonus));
-
-        let bonus = self.slider_on_queen(clr);
-        self.sum(clr, None, None, (60 * bonus, 18 * bonus));
-
-        let bonus = self.knight_on_queen(clr);
-        self.sum(clr, None, None, (16 * bonus, 11 * bonus));
-
-        let bonus = self.restricted(clr).count() as isize;
-        self.sum(clr, None, None, (7 * bonus, 7 * bonus));
-
-        let bonus = self.weak_queen_protection(clr).count() as isize;
-        self.sum(clr, None, None, (14 * bonus, 0));
-
-        self.minor_threat(clr);
-        self.rook_threat(clr);
-    }
-
-    #[inline(always)]
-    fn safe_pawn(&mut self, clr: Color) -> u64 {
-        let bb = (self.pawn_bb(clr) & self.eval.defend_map[clr.idx()])
-            | (self.pawn_bb(clr) & !self.eval.attack_map[clr.opp().idx()]);
-        bb
-    }
-
-    #[inline(always)]
-    fn threat_safe_pawn(&mut self, clr: Color) -> u64 {
-        let bb = self.knight_bb(clr.opp())
-            | self.bishop_bb(clr.opp())
-            | self.rook_bb(clr.opp())
-            | self.queen_bb(clr.opp());
-
-        (bb & get_all_pawn_left_att_mask(self.safe_pawn(clr), clr))
-            | (bb & get_all_pawn_right_att_mask(self.safe_pawn(clr), clr))
-    }
-
-    #[inline(always)]
-    fn weak_enemy(&mut self, clr: Color) -> u64 {
-        let weak_enemy_bb = self.occ_bb(clr.opp())
-            & !get_all_pawn_left_att_mask(self.pawn_bb(clr.opp()), clr.opp())
-            & !get_all_pawn_right_att_mask(self.pawn_bb(clr.opp()), clr.opp())
-            & self.eval.attack_map[clr.idx()];
-        let att_twice = weak_enemy_bb & self.eval.attacked_by_2[clr.idx()];
-        let not_def_twice = weak_enemy_bb & !self.eval.defended_by_2[clr.opp().idx()];
-
-        att_twice | not_def_twice
-    }
-
-    #[inline(always)]
-    fn minor_threat(&mut self, clr: Color) {
-        let bishop = BISHOP + clr;
-        let knight = KNIGHT + clr;
-        let mut bb = (self.weak_enemy(clr) | (self.occ_bb(clr.opp()) & !self.pawn_bb(clr.opp())))
-            & (self.eval.attacked_by[bishop.idx()] | self.eval.attacked_by[knight.idx()]);
-
-        while let Some(sq) = bb.next() {
-            let piece = self.piece_sq(sq);
-            self.sum(clr, Some(sq), Some(piece), MINOR_THREAT[piece.arr_idx()]);
-        }
-    }
-
-    #[inline(always)]
-    fn rook_threat(&mut self, clr: Color) {
-        let piece = ROOK + clr;
-        let mut bb = self.weak_enemy(clr) & self.eval.attacked_by[piece.idx()];
-
-        while let Some(sq) = bb.next() {
-            let piece = self.piece_sq(sq);
-            self.sum(clr, Some(sq), Some(piece), ROOK_THREAT[piece.arr_idx()]);
-        }
-    }
-
-    #[inline(always)]
-    fn hanging(&mut self, clr: Color) -> u64 {
-        let weak_enemy_bb = self.weak_enemy(clr);
-        let att_many =
-            weak_enemy_bb & !self.pawn_bb(clr.opp()) & self.eval.attacked_by_2[clr.idx()];
-        let not_defended = weak_enemy_bb & !self.eval.defend_map[clr.opp().idx()];
-
-        not_defended | att_many
-    }
-
-    #[inline(always)]
-    fn king_threat(&mut self, clr: Color) -> u64 {
-        let king = KING + clr;
-        self.eval.attacked_by[king.idx()] & self.weak_enemy(clr)
-    }
-
-    #[inline(always)]
-    fn pawn_push_threat(&mut self, clr: Color) -> u64 {
-        let clr_push_ranks = [2, 5];
-        let both_occ = self.occ_bb(clr) | self.occ_bb(clr.opp());
-        let mut pawn_threats = 0;
-        let pawn_one_push = get_all_pawn_forward_mask(self.pawn_bb(clr), clr) & !both_occ;
-        let pawn_two_push = get_all_pawn_forward_mask(
-            pawn_one_push & RANK_BITBOARD[clr_push_ranks[clr.idx()]],
-            clr,
-        ) & !both_occ;
-        pawn_threats =
-            (pawn_one_push | pawn_two_push) & !self.eval.attacked_by[(PAWN + clr.opp()).idx()];
-
-        pawn_threats = (pawn_threats & self.eval.attack_map[clr.idx()])
-            | (pawn_threats & !self.eval.attack_map[clr.opp().idx()]);
-
-        return (get_all_pawn_left_att_mask(pawn_threats, clr)
-            | get_all_pawn_right_att_mask(pawn_threats, clr))
-            & self.occ_bb(clr.opp());
-    }
-
-    #[inline(always)]
-    fn slider_on_queen(&mut self, clr: Color) -> isize {
-        if self.queen_bb(clr.opp()).count() != 1 {
-            return 0;
-        }
-
-        // TODO: Try using self.eval.xray here
-        let mut mobility_bb = (self.eval.attacked_by[(QUEEN + clr.opp()).idx()]
-            | self.eval.defended_by[(QUEEN + clr.opp()).idx()])
-            & self.mobility_area(clr);
-
-        mobility_bb = mobility_bb
-            & !self.pawn_bb(clr)
-            & !self.eval.attacked_by[(PAWN + clr.opp()).idx()]
-            & self.eval.attacked_by_2[clr.idx()];
-
-        let diagonal = mobility_bb & self.eval.queen_diagonal[clr.opp().idx()];
-        let orthogonal = mobility_bb & !self.eval.queen_diagonal[clr.opp().idx()];
-
-        let v = if self.queen_bb(clr).count() == 0 { 2 } else { 1 };
-
-        return ((diagonal & self.eval.x_ray[(BISHOP + clr).idx()])
-            | (orthogonal & self.eval.x_ray[(ROOK + clr).idx()]))
-        .count() as isize
-            * v;
-    }
-
-    #[inline(always)]
-    fn knight_on_queen(&mut self, clr: Color) -> isize {
-        if self.queen_bb(clr.opp()).count() != 1 {
-            return 0;
-        }
-
-        let sq = self.queen_bb(clr.opp()).get_lsb();
-        let mut mobility_bb = self.mobility_area(clr)
-            & get_knight_mask(sq, 0, 0, 0)
-            & self.eval.attacked_by[(KNIGHT + clr).idx()];
-
-        mobility_bb =
-            mobility_bb & !self.pawn_bb(clr) & !self.eval.attacked_by[(PAWN + clr.opp()).idx()];
-
-        mobility_bb = mobility_bb
-            & (self.eval.attacked_by_2[clr.idx()] | !self.eval.attacked_by_2[clr.opp().idx()]);
-
-        let v = if self.queen_bb(clr).count() == 0 { 2 } else { 1 };
-
-        return mobility_bb.count() as isize * v;
-    }
-
-    #[inline(always)]
-    fn restricted(&mut self, clr: Color) -> u64 {
-        let restricted_bb = (self.eval.attack_map[clr.idx()] | self.eval.defend_map[clr.idx()])
-            & (self.eval.attack_map[clr.opp().idx()] | self.eval.defend_map[clr.opp().idx()])
-            & !self.eval.attacked_by[(PAWN + clr.opp()).idx()]
-            & !self.eval.defended_by[(PAWN + clr.opp()).idx()]
-            & !((self.eval.attacked_by_2[clr.opp().idx()]
-                | self.eval.defended_by_2[clr.opp().idx()])
-                & (!(self.eval.attacked_by_2[clr.idx()] | self.eval.defended_by_2[clr.idx()])
-                    & (self.eval.attack_map[clr.idx()] | self.eval.defend_map[clr.idx()])));
-
-        restricted_bb
-    }
-
-    #[inline(always)]
-    fn weak_queen_protection(&mut self, clr: Color) -> u64 {
-        let weak_enemy_bb = self.weak_enemy(clr);
-        let mut queen_protect = 0;
-
-        let mut bb = self.queen_bb(clr.opp());
-        while let Some(sq) = bb.next() {
-            queen_protect = weak_enemy_bb & self.get_mask(QUEEN + clr.opp(), sq);
-        }
-
-        return queen_protect;
-    }
-
-    // ************************************************
-    //               9. SPACE EVALUATION              *
-    // ************************************************
-
-    #[inline(always)] // 9. Space
-    fn space(&mut self, clr: Color) {
-        if self.non_pawn_material_eval(clr) + self.non_pawn_material_eval(clr.opp()) < 12222 {
-            return;
-        }
-
-        let own_pawns_blocked =
-            get_all_pawn_forward_mask(self.pawn_bb(clr), clr) & self.pawn_bb(clr.opp());
-        let enemy_pawns_blocked =
-            get_all_pawn_forward_mask(self.pawn_bb(clr.opp()), clr.opp()) & self.pawn_bb(clr);
-        let own_sq_blocked = get_all_pawn_left_att_mask(self.pawn_bb(clr.opp()), clr.opp())
-            & get_all_pawn_right_att_mask(self.pawn_bb(clr.opp()), clr.opp())
-            & get_all_pawn_forward_mask(self.pawn_bb(clr), clr);
-        let enemy_sq_blocked = get_all_pawn_left_att_mask(self.pawn_bb(clr), clr)
-            & get_all_pawn_right_att_mask(self.pawn_bb(clr), clr)
-            & get_all_pawn_forward_mask(self.pawn_bb(clr.opp()), clr.opp());
-
-        let blocked =
-            (own_pawns_blocked | enemy_pawns_blocked | own_sq_blocked | enemy_sq_blocked).count();
-
-        let weight = (self.bb(clr).count() - 3 + blocked.min(9)) as isize;
-
-        let bonus = self.space_area(clr) as isize * weight * weight / 16;
-        self.sum(clr, None, None, (bonus, 0));
-    }
-
-    #[inline(always)]
-    fn space_area(&mut self, clr: Color) -> usize {
-        let mut cnt = 0;
-        let own_pawns_bb = self.pawn_bb(clr);
-        let pawn_behind_bb = self.eval.pawn_behind_masks[clr.idx()];
-        let opp_att_bb = self.eval.attack_map[clr.opp().idx()];
-        let opp_pawn_att_bb = self.eval.attacked_by[(PAWN + clr.opp()).idx()];
-
-        cnt += (CLR_CENTER[clr.idx()] & !opp_pawn_att_bb & !own_pawns_bb).count();
-        cnt += (pawn_behind_bb & CLR_CENTER[clr.idx()] & !opp_att_bb & !own_pawns_bb).count();
-        cnt
-    }
-
-    #[inline(always)]
-    fn king_dist(&mut self, clr: Color, sq: usize) -> usize {
-        let (sq_rank, sq_file) = (get_rank(sq), get_file(sq));
-        let (king_rank, king_file) = (get_rank(self.king_sq(clr)), get_file(self.king_sq(clr)));
-        return (king_rank.abs_diff(sq_rank)).max(king_file.abs_diff(sq_file));
-    }
-
-    #[inline(always)]
-    fn king_ring(&mut self, clr: Color) -> u64 {
-        return KING_RING[self.king_sq(clr)] & !get_pawn_2_att(self.pawn_bb(clr), clr);
-    }
-
-    #[inline(always)]
-    fn opp_color_bishops(&mut self, clr: Color) -> bool {
-        let clr_bishop = self.bishop_bb(clr).count();
-        let opp_clr_bishop = self.bishop_bb(clr.opp()).count();
-
-        return clr_bishop == 1
-            && opp_clr_bishop == 1
-            && has_bishop_pair(self.bishop_bb(clr) | self.bishop_bb(clr.opp()));
-    }
-
-    #[inline(always)]
-    fn passed_pawn(&mut self, clr: Color) {
-        let piece = PAWN + clr;
-
-        let mut bb = self.pawn_bb(clr);
-        while let Some(sq) = bb.next() {
-            if !self.passed_leverable(sq, clr) {
-                continue;
-            }
-            let king_proximity = self.king_proximity(sq, clr);
-
-            let passed_block = self.passed_blocked(sq, clr);
-
-            let passed_file = self.passed_file(sq);
-            self.sum(clr, Some(sq), Some(piece), (0, king_proximity));
-            self.sum(clr, Some(sq), Some(piece), PASSED_PAWN_REW[clr.idx()][get_rank(sq)]);
-            self.sum(clr, Some(sq), Some(piece), (passed_block, passed_block));
-            self.sum(clr, Some(sq), Some(piece), (-11 * passed_file, -8 * passed_file));
-        }
-    }
-
-    #[inline(always)]
-    fn passed_leverable(&mut self, sq: usize, clr: Color) -> bool {
-        if !self.candidate_passed(sq, clr) {
-            // println!("Candidate Not Passed on sq: {:?}", sq);
-            return false;
-        }
-
-        if !self.blocked_pawn(sq, clr, self.pawn_bb(clr.opp())) {
-            return true;
-        }
-
-        let mut bb = PAWN_ATTACK_LOOKUP[clr.opp().idx()][sq] & self.pawn_bb(clr);
-
-        while let Some(square) = bb.next() {
-            let front_sq = self.front_sq(square, clr);
-            let is_occupied = self.occ_bb(clr.opp()).is_set(front_sq);
-            let is_more_def = self.eval.attack_map[clr.idx()].is_set(front_sq)
-                || !self.eval.attacked_by_2[clr.opp().idx()].is_set(front_sq);
-            if !is_occupied && is_more_def {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    #[inline(always)]
-    fn passed_file(&mut self, sq: usize) -> isize {
-        let file = get_file(sq) as isize;
-        file.min(7 - file)
-    }
-
-    #[inline(always)]
-    fn passed_blocked(&mut self, sq: usize, clr: Color) -> isize {
-        let (own, enemy) = self.both_occ_bb(clr);
-        let clr_rank = CLR_RANK[clr.idx()][get_rank(sq) as usize];
-
-        if clr_rank <= 2 || (own | enemy).is_set(self.front_sq(sq, clr)) {
-            return 0;
-        }
-
-        let weight = 5 * clr_rank - 13;
-        let forward = PAWN_FORWARD_SPANS[clr.idx()][sq];
-        let backward = PAWN_FORWARD_SPANS[clr.opp().idx()][sq];
-        let forward_lr = FORWARD_SPANS_LR[clr.idx()][sq];
-
-        let mut defended_bb =
-            forward & (self.eval.defend_map[clr.idx()] | self.eval.attack_map[clr.idx()]);
-
-        // print_bitboard(forward, None);
-        // print_bitboard(self.eval.defend_map[clr.opp().idx()], None);
-        // print_bitboard(self.eval.attack_map[clr.opp().idx()], None);
-
-        let mut unsafe_bb = forward
-            & (self.eval.defend_map[clr.opp().idx()] | self.eval.attack_map[clr.opp().idx()]);
-
-        // print_bitboard(unsafe_bb, None);
-
-        let mut wunsafe_bb = forward_lr
-            & (self.eval.defend_map[clr.opp().idx()] | self.eval.attack_map[clr.opp().idx()]);
-
-        let mut is_defended1 = defended_bb.is_set(self.front_sq(sq, clr));
-
-        let mut is_unsafe1 = unsafe_bb.is_set(self.front_sq(sq, clr));
-
-        if (self.queen_bb(clr) | self.rook_bb(clr)) & backward != 0 {
-            is_defended1 = true;
-            defended_bb = 1;
-        }
-
-        if (self.queen_bb(clr.opp()) | self.rook_bb(clr.opp())) & backward != 0 {
-            is_unsafe1 = true;
-            unsafe_bb = 1;
-        }
-
-        let mut k = 0;
-
-        // println!("Unsafe: {:?}", unsafe_bb.count());
-        // println!("Wunsafe: {:?}", wunsafe_bb.count());
-        // println!("is_unsafe1: {:?}", is_unsafe1);
-        // println!("is_defended1: {:?}", is_defended1);
-
-        if unsafe_bb == 0 && wunsafe_bb == 0 {
-            k = 35;
-        } else if unsafe_bb == 0 {
-            k = 20;
-        } else if !is_unsafe1 {
-            k = 9;
-        }
-
-        if is_defended1 {
-            k += 5;
-        }
-
-        return k * (weight as isize);
-    }
-
-    #[inline(always)]
-    fn king_proximity(&mut self, sq: usize, clr: Color) -> isize {
-        let mut score = 0;
-
-        let clr_rank = CLR_RANK[clr.idx()][get_rank(sq) as usize];
-
-        if clr_rank <= 2 {
-            return 0;
-        }
-
-        let weight = (5 * clr_rank - 13) as isize;
-
-        let front_sq = self.front_sq(sq, clr);
-
-        score += (self.king_dist(clr.opp(), front_sq).min(5) * 19 / 4) as isize * weight;
-        score -= (self.king_dist(clr, front_sq).min(5) * 2) as isize * weight;
-
-        // Consider another push if the next square is the queening square
-        if clr_rank != 6 {
-            score -= self.king_dist(clr, front_sq).min(5) as isize * weight;
-        }
-        score
-    }
-
-    #[inline(always)]
-    fn candidate_passed(&mut self, sq: usize, clr: Color) -> bool {
-        let forward = PAWN_FORWARD_SPANS[clr.idx()][sq];
-        let forward_lr = FORWARD_SPANS_LR[clr.idx()][sq];
-        let our_pawns = self.pawn_bb(clr);
-        let their_pawns = self.pawn_bb(clr.opp());
-
-        // Own pawn ahead? Blocked by same-file pawn
-        // println!("Try Candidate Passed : {:?}", sq);
-        if forward & our_pawns != 0 {
-            return false;
-        }
-
-        // No enemy pawn in any of the 3 forward files → clearly candidate
-        if forward & their_pawns == 0 && forward_lr & their_pawns == 0 {
-            return true;
-        }
-
-        let one_forward = self.front_sq(sq, clr);
-        let double_forward = self.front_sq(one_forward, clr);
-        if (double_forward < 64 && FORWARD_SPANS_LR[clr.idx()][double_forward] & their_pawns != 0)
-            || PAWN_FORWARD_SPANS[clr.idx()][self.front_sq(sq, clr)] & their_pawns != 0
-        {
-            return false;
-        }
-
-        if FORWARD_SPANS_LR[clr.idx()][self.back_sq(sq, clr)] & their_pawns == 0
-            && self.blocked_pawn(sq, clr, their_pawns)
-            && CLR_RANK[clr.idx()][get_rank(sq)] > 3
-        {
-            let mut bb = PAWN_ATTACK_LOOKUP[clr.opp().idx()][sq] & self.pawn_bb(clr);
-            while let Some(square) = bb.next() {
-                let front_sq = self.front_sq(square, clr);
-                let double_front_sq = self.front_sq(front_sq, clr);
-                if !their_pawns.is_set(front_sq) && !their_pawns.is_set(double_front_sq) {
-                    return true;
-                }
-            }
-        }
-
-        if self.blocked_pawn(sq, clr, their_pawns) {
-            return false;
-        }
-
-        let lever_mask = PAWN_ATTACK_LOOKUP[clr.idx()][sq] & their_pawns;
-        let leverpush_mask = PAWN_ATTACK_LOOKUP[clr.idx()][self.front_sq(sq, clr)] & their_pawns;
-        let phalanx_mask = PAWN_ATTACK_LOOKUP[clr.idx()][self.back_sq(sq, clr)] & our_pawns;
-
-        let lever = lever_mask.count();
-        let leverpush = leverpush_mask.count();
-        let phalanx = phalanx_mask.count();
-        let supported = self.supported_pawn(sq, clr) as usize;
-
-        if lever > supported + 1 {
-            return false;
-        }
-        if leverpush > phalanx {
-            return false;
-        }
-        if lever > 0 && leverpush > 0 {
-            return false;
-        }
-
-        true
-    }
-
-    // ************************************************
-    //                   11. KING                     *
-    // ************************************************
-
-    #[inline(always)]
-    fn king_eval(&mut self, clr: Color) {
-        let king_sq = self.king_sq(clr);
-
-        let bonus = self.king_danger(clr);
-        self.sum(clr, None, None, ((bonus * bonus) / 4096, bonus / 16));
-
-        let bonus = self.eval.king_shelter[clr.idx()];
-        self.sum(clr, None, None, (-bonus.0, 0)); // Shelter Strength
-        self.sum(clr, None, None, (bonus.1, 0)); // Shelter Storm
-
-        let bonus = if self.pawnless_flank(king_sq, clr) { 1 } else { 0 };
-        self.sum(clr, None, None, (17 * bonus, 95 * bonus));
-
-        let bonus = self.flank_attack(clr);
-        self.sum(clr, None, None, (8 * bonus, 0));
-
-        let bonus = self.eval.king_pawn_dx[clr.idx()] as isize;
-        self.sum(clr, None, None, (0, -16 * bonus));
-
-        // FIXME: This is not correct, the function is wrong
-        let bonus = self.endgame_shelter(clr);
-        self.sum(clr, None, None, (0, bonus));
-    }
-
-    #[inline(always)]
-    fn king_danger(&mut self, clr: Color) -> isize {
-        let count = self.king_attackers_count(clr);
-        // println!("King Attackers Count: {:?}", count);
-
-        let weight = self.king_attackers_weight(clr);
-        // println!("King Attackers Weight: {:?}", weight);
-
-        let king_att = self.king_attacks(clr);
-        // println!("King Attacks: {:?}", king_att);
-
-        let weak = self.weak_bonus(clr).count() as isize;
-        // print_bitboard(self.eval.king_ring[clr.opp().idx()], None);
-        // print_bitboard(self.weak_squares(clr), None);
-        // print_bitboard(self.weak_bonus(clr), None);
-        // println!("Weak Bonus: {:?}", weak);
-
-        let unsafe_checks = self.unsafe_checks(clr).count() as isize;
-        // println!("Unsafe Checks: {:?}", unsafe_checks);
-        // print_bitboard(self.unsafe_checks(clr), None);
-
-        let flank_att = self.flank_attack(clr);
-        // println!("Flank Attack: {:?}", flank_att);
-
-        let flank_def = self.flank_defense(clr);
-        // println!("Flank defense: {:?}", flank_def);
-
-        let no_queen = if self.queen_bb(clr).count() > 0 { 0 } else { 1 };
-        // println!("No Queen: {:?}", no_queen);
-
-        let knight_defender = if self.knight_defender(clr.opp()).count() > 0 { 1 } else { 0 };
-        // println!("Knight Defender: {:?}", knight_defender);
-
-        // let knight_safe = self.safe_check(clr, KNIGHT + clr);
-        // println!("Knight Safe Check");
-        // print_bitboard(knight_safe, None);
-
-        // let rook_safe = self.safe_check(clr, ROOK + clr);
-        // println!("Rook Safe Check");
-        // print_bitboard(rook_safe, None);
-
-        // let bishop_safe = self.safe_check(clr, BISHOP + clr);
-        // println!("Bishop Safe Check");
-        // print_bitboard(bishop_safe, None);
-
-        // let queen_safe = self.safe_check(clr, QUEEN + clr);
-        // println!("Queen Safe Check");
-        // print_bitboard(queen_safe, None);
-
-        let v = count * weight + 69 * king_att + 185 * weak - 100 * knight_defender
-            + 148 * unsafe_checks
-            - 4 * flank_def
-            + (3 * flank_att * flank_att / 8)
-            - 873 * no_queen
-            - (6 * (self.eval.king_shelter[clr.idx()].0 - self.eval.king_shelter[clr.idx()].1) / 8) //self.shelter(clr).0 - self.shelter(clr).1
-            + self.eval.mobility[clr.idx()]
-            - self.eval.mobility[clr.opp().idx()]
-            + 37
-            + 772 * (self.safe_check(clr, QUEEN + clr).count() as f64).min(1.45) as isize
-            + 1084 * (self.safe_check(clr, ROOK + clr).count() as f64).min(1.75) as isize
-            + 645 * (self.safe_check(clr, BISHOP + clr).count() as f64).min(1.50) as isize
-            + 792 * (self.safe_check(clr, KNIGHT + clr).count() as f64).min(1.62) as isize;
-        // println!("V Score: {:?}", v);
-        // println!("-------------------------------");
-
-        if v > 100 {
-            return v;
-        };
-        return 0;
-    }
-
-    #[inline(always)]
-    fn flank_defense(&mut self, clr: Color) -> isize {
-        let king_sq = self.king_sq(clr.opp());
-        let flanks = ISOLATED_PAWN_LOOKUP[king_sq]
-            | FILE_BITBOARD[get_file(king_sq)]
-            | FILE_BITBOARD[FLANK_ADDITIONAL_FILE[get_file(king_sq)]];
-
-        let att_1 = flanks
-            & FLANK_MASK[clr.opp().idx()]
-            & (self.eval.attack_map[clr.opp().idx()] | self.eval.defend_map[clr.opp().idx()]);
-
-        // println!("Attack 1");
-        // print_bitboard(att_1, None);
-
-        return att_1.count() as isize;
-        // self.eval.sum(clr, None, None, (count as isize * 2, count as isize * 2));
-    }
-
-    #[inline(always)]
-    fn flank_attack(&mut self, clr: Color) -> isize {
-        let king_sq = self.king_sq(clr.opp());
-        let flanks = ISOLATED_PAWN_LOOKUP[king_sq]
-            | FILE_BITBOARD[get_file(king_sq)]
-            | FILE_BITBOARD[FLANK_ADDITIONAL_FILE[get_file(king_sq)]];
-
-        let att_1 = flanks
-            & FLANK_MASK[clr.opp().idx()]
-            & (self.eval.attack_map[clr.idx()] | self.eval.defend_map[clr.idx()]);
-
-        let att_2 = flanks
-            & FLANK_MASK[clr.opp().idx()]
-            & (self.eval.attacked_by_2[clr.idx()] | self.eval.defended_by_2[clr.idx()]);
-
-        // println!("Attack 1");
-        // print_bitboard(att_1, None);
-
-        // println!("Attack 2");
-        // print_bitboard(att_2, None);
-
-        return (att_1.count() + att_2.count()) as isize;
-    }
-
-    #[inline(always)]
-    fn king_blockers(&mut self, clr: Color) {
-        todo!()
-    }
-
-    #[inline(always)]
-    fn endgame_shelter(&mut self, clr: Color) -> isize {
-        self.eval.king_shelter[clr.idx()].2
-    }
-
-    #[inline(always)]
-    fn knight_defender(&mut self, clr: Color) -> u64 {
-        (self.eval.attacked_by[(KNIGHT + clr).idx()] | self.eval.defended_by[(KNIGHT + clr).idx()])
-            & (self.eval.attacked_by[(KING + clr).idx()]
-                | self.eval.defended_by[(KING + clr).idx()])
-    }
-
-    #[inline(always)]
-    fn unsafe_checks(&mut self, clr: Color) -> u64 {
-        let knight_unsafe =
-            self.eval.checks[(KNIGHT + clr).idx()] & !self.safe_check(clr, KNIGHT + clr);
-
-        let bishop_unsafe =
-            self.eval.checks[(BISHOP + clr).idx()] & !self.safe_check(clr, BISHOP + clr);
-
-        let rook_unsafe = self.eval.checks[(ROOK + clr).idx()] & !self.safe_check(clr, ROOK + clr);
-
-        return knight_unsafe | bishop_unsafe | rook_unsafe;
-    }
-
-    #[inline(always)]
-    fn safe_check(&mut self, clr: Color, piece: Piece) -> u64 {
-        let checks = match piece.kind() {
-            PAWN => 0,
-            KNIGHT => self.eval.checks[piece.idx()],
-            KING => 0,
-            BISHOP => self.eval.checks[piece.idx()] & !self.eval.checks[(QUEEN + clr).idx()],
-            ROOK => self.eval.checks[piece.idx()],
-            QUEEN => {
-                self.eval.checks[piece.idx()]
-                    & !self.eval.checks[(ROOK + clr).idx()]
-                    & !(self.eval.attacked_by[(QUEEN + clr.opp()).idx()]
-                        | self.eval.defended_by[(QUEEN + clr.opp()).idx()])
-            }
-            _ => panic!("There is other peace that was not expected here"),
-        };
-
-        let weak_squares = self.weak_squares(clr) & self.eval.attacked_by_2[clr.idx()];
-
-        return !self.occ_bb(clr)
-            & (!(self.eval.attack_map[clr.opp().idx()] | self.eval.defend_map[clr.opp().idx()])
-                | weak_squares)
-            & checks;
-    }
-
-    #[inline(always)]
-    fn weak_squares(&mut self, clr: Color) -> u64 {
-        let enemy_att_2 =
-            self.eval.attacked_by_2[clr.opp().idx()] | self.eval.defended_by_2[clr.opp().idx()];
-
-        let not_att_2_times =
-            (self.eval.attack_map[clr.idx()] | self.eval.defend_map[clr.idx()]) & !enemy_att_2;
-
-        (self.eval.attack_map[clr.idx()] | self.eval.defend_map[clr.idx()])
-            & (not_att_2_times
-                & !(self.eval.attack_map[clr.opp().idx()] | self.eval.defend_map[clr.opp().idx()]))
-            | (not_att_2_times
-                & (self.eval.attacked_by[(KING + clr.opp()).idx()]
-                    | self.eval.defended_by[(KING + clr.opp()).idx()]
-                    | self.eval.attacked_by[(QUEEN + clr.opp()).idx()]
-                    | self.eval.defended_by[(QUEEN + clr.opp()).idx()]))
-    }
-
-    #[inline(always)]
-    fn weak_bonus(&mut self, clr: Color) -> u64 {
-        self.weak_squares(clr) & self.eval.king_ring[clr.opp().idx()]
-    }
-
-    #[inline(always)]
-    fn king_attacks(&mut self, clr: Color) -> isize {
-        self.eval.king_att[clr.idx()] as isize
-    }
-
-    #[inline(always)]
-    fn king_attackers_weight(&mut self, clr: Color) -> isize {
-        self.eval.king_att_weight[clr.idx()]
-    }
-
-    #[inline(always)]
-    fn king_attackers_count(&mut self, clr: Color) -> isize {
-        self.eval.king_att_count[clr.idx()] as isize
-    }
-
-    #[inline(always)]
-    fn check(&mut self, clr: Color) {
-        let king_sq = self.king_sq(clr.opp());
-        self.eval.checks[(KNIGHT + clr).idx()] = (self.eval.attacked_by[(KNIGHT + clr).idx()]
-            | self.eval.defended_by[(KNIGHT + clr).idx()])
-            & self.x_ray_mask(KNIGHT + clr.opp(), king_sq);
-
-        self.eval.checks[(BISHOP + clr).idx()] = (self.eval.attacked_by[(BISHOP + clr).idx()]
-            | self.eval.defended_by[(BISHOP + clr).idx()])
-            & self.x_ray_mask(BISHOP + clr.opp(), king_sq);
-
-        self.eval.checks[(ROOK + clr).idx()] = (self.eval.attacked_by[(ROOK + clr).idx()]
-            | self.eval.defended_by[(ROOK + clr).idx()])
-            & self.x_ray_mask(ROOK + clr.opp(), king_sq);
-
-        self.eval.checks[(QUEEN + clr).idx()] = (self.eval.attacked_by[(QUEEN + clr).idx()]
-            | self.eval.defended_by[(QUEEN + clr).idx()])
-            & self.x_ray_mask(QUEEN + clr.opp(), king_sq);
-    }
-
-    #[inline(always)]
-    fn shelter(&mut self, clr: Color) -> (isize, isize, isize) {
-        let king_sq = self.king_sq(clr.opp());
-        let mut king_strenght = self.strength_square(king_sq, clr);
-        let mut king_storm = self.storm_square(king_sq, clr);
-        // println!("Storm SQ king{:?}", self.storm_square(king_sq, clr));
-
-        if self.castling().short(clr.opp()) != 0 {
-            let short_castle_sq = if clr.opp().is_white() { 6 } else { 62 };
-            let short_castle_strength = self.strength_square(short_castle_sq, clr);
-            let short_castle_storm = self.storm_square(short_castle_sq, clr);
-            // println!("Storm SQ short{:?}", self.storm_square(short_castle_sq, clr));
-
-            if (short_castle_storm.0 - short_castle_strength) < (king_storm.0 - king_strenght) {
-                king_strenght = short_castle_strength;
-                king_storm = short_castle_storm;
-            }
-        }
-
-        if self.castling().long(clr.opp()) != 0 {
-            let long_castle_sq = if clr.opp().is_white() { 2 } else { 58 };
-            let long_castle_strength = self.strength_square(long_castle_sq, clr);
-            let long_castle_storm = self.storm_square(long_castle_sq, clr);
-            // println!("Storm SQ long{:?}", self.storm_square(long_castle_sq, clr));
-
-            if (long_castle_storm.0 - long_castle_strength) < (king_storm.0 - king_strenght) {
-                king_strenght = long_castle_strength;
-                king_storm = long_castle_storm;
-            }
-        }
-
-        return (king_strenght, king_storm.0, king_storm.1);
-    }
-
-    #[inline(always)]
-    fn storm_square(&mut self, sq: usize, clr: Color) -> (isize, isize) {
-        let mut v = 0;
-        let mut ev = 5;
-
-        let file = get_file(sq);
-        let sq = match file {
-            0 => sq + 1,
-            7 => sq - 1,
-            _ => sq,
-        };
-
-        for square in (sq - 1)..(sq + 2) {
-            // FIXME: ALL Squares forward ????????????
-            let us_bb: u64 = (PAWN_FORWARD_SPANS[clr.opp().idx()][square] | Bitboard::init(square))
-                & (self.pawn_bb(clr.opp()) & !self.eval.attacked_by[(PAWN + clr).idx()]);
-
-            let them_bb: u64 = (PAWN_FORWARD_SPANS[clr.opp().idx()][square]
-                | Bitboard::init(square))
-                & self.pawn_bb(clr);
-
-            let mut us = 0;
-            let mut them = 0;
-
-            if us_bb != 0 {
-                us = get_rank(if clr.is_white() { us_bb.get_msb() } else { us_bb.get_lsb() });
-            }
-
-            if them_bb != 0 {
-                them = get_rank(if clr.is_white() { them_bb.get_msb() } else { them_bb.get_lsb() });
-            }
-
-            // println!("Sq: {:?}, Square: {:?}, Us: {:?}, Them: {:?}", sq, square, us, them);
-            if us > 0 && (them as isize) == (us as isize) - clr.sign() {
-                // v += BLOCKED_STORM[0][CLR_RANK[clr.idx()][them]];
-                // ev += BLOCKED_STORM[1][CLR_RANK[clr.idx()][them]];
-
-                if them == 0 {
-                    v += BLOCKED_STORM[0][0];
-                    ev += BLOCKED_STORM[1][0];
-                } else {
-                    v += BLOCKED_STORM[0][CLR_RANK[clr.idx()][them]];
-                    ev += BLOCKED_STORM[1][CLR_RANK[clr.idx()][them]];
-                }
-            } else {
-                // println!("First: {:?}", get_rank(square));
-                // println!("Them: {:?}", them);
-                // println!("CLR_RANK: {:?}", CLR_RANK[clr.idx()]);
-                // println!("GET Rank{:?}", get_rank(them));
-                if them == 0 {
-                    v += UNBLOCKED_STORM[get_file(square)][0];
-                } else {
-                    v += UNBLOCKED_STORM[get_file(square)][CLR_RANK[clr.idx()][them]];
-                }
-            }
-        }
-        return (v, ev);
-    }
-
-    #[inline(always)]
-    fn strength_square(&mut self, sq: usize, clr: Color) -> isize {
-        let mut score = 5;
-
-        let file = get_file(sq);
-        let sq = match file {
-            0 => sq + 1,
-            7 => sq - 1,
-            _ => sq,
-        };
-
-        for square in (sq - 1)..(sq + 2) {
-            let mut us = 0;
-            let us_bb: u64 = (PAWN_FORWARD_SPANS[clr.opp().idx()][square] | Bitboard::init(square))
-                & (self.pawn_bb(clr.opp()) & !self.eval.attacked_by[(PAWN + clr).idx()]);
-
-            if us_bb != 0 {
-                us = get_rank(if clr.is_white() { us_bb.get_msb() } else { us_bb.get_lsb() });
-            }
-
-            if us == 0 {
-                score += WEAKNESS[get_file(square)][0];
-            } else {
-                score += WEAKNESS[get_file(square)][CLR_RANK[clr.idx()][us]];
-            }
-        }
-
-        return score;
-    }
-
-    #[inline(always)]
-    fn pawnless_flank(&mut self, sq: usize, clr: Color) -> bool {
-        let all_pawns = self.pawn_bb(clr) | self.pawn_bb(clr.opp());
-        let file = get_file(sq);
-        let flanks = ISOLATED_PAWN_LOOKUP[sq]
-            | FILE_BITBOARD[file]
-            | FILE_BITBOARD[FLANK_ADDITIONAL_FILE[file]];
-
-        return (flanks & all_pawns) == 0;
-    }
-
-    // ************************************************
-    //                  11. TEMPO                     *
-    // ************************************************
-
-    #[inline(always)]
-    fn tempo(&mut self, clr: Color) {
-        let bonus = (TEMPO_WT, TEMPO_WT);
-        self.sum(clr, None, None, bonus);
-    }
 }
 
 #[cfg(test)]
 mod tests {
 
-    use crate::engine::board::{self, fen::FenTrait};
+    use crate::engine::board::fen::FenTrait;
+    use crate::engine::evaluation::test_evaluation::SF_EVAL;
 
     use super::*;
-
-    struct SFEval<'a> {
-        fen: &'a str,
-        phase: isize,
-        eval: isize,
-
-        material: isize,
-        psqt: isize,
-        imbalance: isize,
-        pawns: isize,
-        piece: isize,
-        mobility: isize,
-        threats: isize,
-        passed_pawn: isize,
-        space: isize,
-        king: isize,
-        winnable: isize,
-        tempo: isize,
-    }
-
-    const SF_EVAL: [SFEval; 11] = [
-        SFEval {
-            fen: "r3r1k1/3q1pp1/p2pb2p/Np6/1P1QPn2/5N1P/1P3PP1/R3R1K1 w - - 0 0",
-            phase: 106,
-            eval: -34,
-
-            material: -46,
-            psqt: -56,
-            imbalance: 36,
-            pawns: 6,
-            piece: 52,
-            mobility: -30,
-            threats: 4,
-            passed_pawn: 0,
-            space: 12,
-            king: -67,
-            winnable: -123465, // FIXME:
-            tempo: 28,
-        },
-        SFEval {
-            fen: "1rb1r1k1/2q2pp1/1b1p2np/1pp5/3Pn3/1B2BNNP/1P1Q1PP1/R3R1K1 w - - 0 0",
-            phase: 128,
-            eval: -244,
-
-            material: -124,
-            psqt: -2,
-            imbalance: -50,
-            pawns: -91,
-            piece: 55,
-            mobility: 18,
-            threats: -127,
-            passed_pawn: 0,
-            space: -21,
-            king: 39,
-            winnable: -123465, // FIXME:
-            tempo: 28,
-        },
-        SFEval {
-            fen: "3r2k1/2p2bpp/p2r4/P2PpP2/BR1q4/7P/5PP1/2R1Q1K1 b - - 0 0",
-            phase: 89,
-            eval: 404,
-
-            material: 148,
-            psqt: -50,
-            imbalance: 19,
-            pawns: 40,
-            piece: 40,
-            mobility: 61,
-            threats: 165,
-            passed_pawn: 0,
-            space: 0,
-            king: 102,
-            winnable: -123465, // FIXME:
-            tempo: -28,
-        },
-        SFEval {
-            fen: "rnb1k2r/2p1ppPp/5bn1/p1p5/P2p2p1/P2P1P1P/6P1/RNB1KBNR b KQkq - 0 0",
-            phase: 89,
-            eval: -444,
-
-            material: -151,
-            psqt: -90,
-            imbalance: -65,
-            pawns: -25,
-            piece: 7,
-            mobility: -124,
-            threats: -71,
-            passed_pawn: 233,
-            space: 0,
-            king: -124,
-            winnable: -123465, // FIXME:
-            tempo: -28,
-        },
-        SFEval {
-            fen: "r1bqk1r1/1p1p1n2/p1n2pN1/2p1b2Q/2P1Pp2/1PN5/PB4PP/R4RK1 w q - 0 0",
-            phase: 128,
-            eval: -836,
-
-            material: -825,
-            psqt: 154,
-            imbalance: -148,
-            king: -135,
-            mobility: 103,
-            passed_pawn: 10,
-            pawns: 93,
-            piece: -11,
-            space: -15,
-            threats: -104,
-            winnable: -123465, // FIXME:
-            tempo: 28,
-        },
-        SFEval {
-            fen: "r1n2N1k/2n2K1p/3pp3/5Pp1/b5R1/8/1PPP4/8 w - - 0 0",
-            phase: 20,
-            eval: -1476,
-
-            material: -1743,
-            psqt: 124,
-            imbalance: -123,
-            king: -187,
-            mobility: -21,
-            passed_pawn: 231,
-            pawns: -46,
-            piece: 159,
-            space: 0,
-            threats: 154,
-            winnable: -64,
-            tempo: 28,
-        },
-        SFEval {
-            fen: "4rrk1/Rpp3pp/6q1/2PPn3/4p3/2N5/1P2QPPP/5RK1 w - - 0 0",
-            phase: 88,
-            eval: 204,
-
-            material: 149,
-            psqt: -35,
-            imbalance: 29,
-            king: -204,
-            mobility: -25,
-            passed_pawn: 67,
-            pawns: 143,
-            piece: -19,
-            space: 0,
-            threats: 91,
-            winnable: -3,
-            tempo: 28,
-        },
-        SFEval {
-            fen: "r3kb1r/3n1ppp/p3p3/1p1pP2P/P3PBP1/4P3/1q2B3/R2Q1K1R b kq - 0 0",
-            phase: 107,
-            eval: -348,
-
-            material: -90,
-            psqt: 151,
-            imbalance: -21,
-            king: -102,
-            mobility: 25,
-            passed_pawn: -38,
-            pawns: -169,
-            piece: -45,
-            space: -3,
-            threats: -24,
-            winnable: -3,
-            tempo: -28,
-        },
-        SFEval {
-            fen: "rnb2rk1/pp2q2p/3p4/2pP2p1/2P1Pp2/2N5/PP1QBRPP/R5K1 w - - 0 0",
-            phase: 106,
-            eval: 172,
-
-            material: 0,
-            psqt: 137,
-            imbalance: 0,
-            king: -30,
-            mobility: 68,
-            passed_pawn: 0,
-            pawns: -39,
-            piece: 4,
-            space: 20,
-            threats: -14,
-            winnable: 4,
-            tempo: 28,
-        },
-        SFEval {
-            fen: "8/2P1P3/b1B2p2/1pPRp3/2k3P1/P4pK1/nP3p1p/N7 w - - 0 0",
-            phase: 106,
-            eval: 1612,
-
-            material: 1375,
-            psqt: -82,
-            imbalance: 10,
-            king: 143,
-            mobility: 165,
-            passed_pawn: -336,
-            pawns: 66,
-            piece: -14,
-            space: 0,
-            threats: 117,
-            winnable: 143,
-            tempo: 28,
-        },
-        SFEval {
-            fen: "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1",
-            phase: 128,
-            eval: 28,
-
-            material: 0,
-            psqt: 0,
-            imbalance: 0,
-            king: 0,
-            mobility: 0,
-            passed_pawn: 0,
-            pawns: 0,
-            piece: 0,
-            space: 0,
-            threats: 0,
-            winnable: 0,
-            tempo: 28,
-        },
-    ];
-
-    // Calculating Stockfish evaluation of certain element based on phase
-    // The Stockfish phase goes from 128 to 0
-    // println!("{:?}", stockfish_eval(106, -3, 55));
-    fn stockfish_eval(phase: isize, mg_value: isize, eg_value: isize) -> isize {
-        (phase * mg_value + (128 - phase) * eg_value) / 128
-    }
-
-    // FIXME: DEPRECATE: TEST
-    #[test]
-    fn testing() {
-        let phase: isize = 128;
-        // println!("      material: {:?},", stockfish_eval(phase, -124, -206)); // material
-        // println!("          psqt: {:?},", stockfish_eval(phase, -89, -94)); // psqt
-        // println!("     imbalance: {:?},", -64); // imbalance
-        // println!("         pawns: {:?},", stockfish_eval(phase, 43, 36)); // pawns
-        // println!("         piece: {:?},", stockfish_eval(phase, -16, 54)); // piece
-        // println!("      mobility: {:?},", stockfish_eval(phase, -135, -103)); // mobility
-        // println!("       threats: {:?},", stockfish_eval(phase, -74, -68)); // threats
-        // println!("   passed_pawn: {:?},", stockfish_eval(phase, 236, 229)); // threats
-        // println!("         space: {:?},", stockfish_eval(phase, 0, 0)); // space
-        // println!("          king: {:?},", stockfish_eval(phase, -172, -31)); // king
-        // println!("         tempo: {:?},", -28); // Tempo
-
-        // 107 OR 106 ?
-        println!("      material: {:?},", stockfish_eval(phase, -80, -145)); // material
-        println!("          psqt: {:?},", stockfish_eval(phase, 170, 60)); // psqt
-        println!("      imbalance: {:?},", stockfish_eval(phase, -21, -21)); // imbalance
-        println!("           king: {:?},", stockfish_eval(phase, -75, -31)); // king
-        println!("      mobility: {:?},", stockfish_eval(phase, 17, 71)); // mobility
-        println!("   passed_pawn: {:?},", stockfish_eval(phase, -32, -73)); // passed pawns
-        println!("         pawns: {:?},", stockfish_eval(phase, -171, -162)); // pawns
-        println!("         piece: {:?},", stockfish_eval(phase, 0, 0)); // pieces
-        println!("         space: {:?},", stockfish_eval(phase, 0, 0)); // space
-        println!("       threats: {:?},", stockfish_eval(phase, 0, 0)); // threats
-        println!("      winnable: {:?},", stockfish_eval(phase, 0, 0)); // winnable
-
-        // println!("      material: {:?},", stockfish_eval(phase, 1276, 1380)); // material
-        // println!("          psqt: {:?},", stockfish_eval(phase, 33, -88)); // psqt
-        // println!("      imbalance: {:?},", stockfish_eval(phase, 10, 10)); // imbalance
-        // println!("           king: {:?},", stockfish_eval(phase, 663, 118)); // king
-        // println!("      mobility: {:?},", stockfish_eval(phase, 92, 169)); // mobility
-        // println!("   passed_pawn: {:?},", stockfish_eval(phase, -565, -325)); // passed pawns
-        // println!("         pawns: {:?},", stockfish_eval(phase, 12, 69)); // pawns
-        // println!("         piece: {:?},", stockfish_eval(phase, 65, -18)); // pieces
-        // println!("         space: {:?},", stockfish_eval(phase, 0, 0)); // space
-        // println!("       threats: {:?},", stockfish_eval(phase, 220, 112)); // threats
-        // println!("      winnable: {:?},", stockfish_eval(phase, 0, 151)); // winnable
-    }
-
-    // Because of the devision and things like that  I am using offset of 1 to be a mistake in my calculation
-
-    fn eval_assert(actual: isize, expected: isize) {
-        if actual - expected == 1 {
-            assert_eq!(actual - 1, expected);
-        } else if actual - expected == -1 {
-            assert_eq!(actual + 1, expected);
-        } else {
-            assert_eq!(actual, expected);
-        }
-    }
-
-    // NOTE: 1. MATERIAL [TEST: WORKS]
-    #[test]
-    fn material_test() {
-        for obj in &SF_EVAL {
-            // if obj.fen != "3r2k1/2p2bpp/p2r4/P2PpP2/BR1q4/7P/5PP1/2R1Q1K1 b - - 0 0" {
-            //     continue;
-            // }
-
-            let mut board = Board::read_fen(obj.fen);
-            board.init();
-            board.material_eval(WHITE);
-            board.material_eval(BLACK);
-
-            if board.calculate_score() != obj.material {
-                println!("assertion `{:?} == {:?}` failed", board.calculate_score(), obj.material);
-            } else {
-                println!("assertion `{:?} == {:?}` success", board.calculate_score(), obj.material);
-                assert_eq!(board.calculate_score(), obj.material);
-            }
-
-            // board.calculate_score();
-            // assert_eq!(board.calculate_score(), obj.material);
-        }
-    }
 
     // NOTE: 2. PSQT [TEST: WORKS]
     #[test]
@@ -2486,14 +237,13 @@ mod tests {
             board.psqt_eval(WHITE);
             board.psqt_eval(BLACK);
 
-            if board.calculate_score() != obj.psqt {
-                println!("assertion `{:?} == {:?}` failed", board.calculate_score(), obj.psqt);
-            } else {
-                println!("assertion `{:?} == {:?}` success", board.calculate_score(), obj.psqt);
-                assert_eq!(board.calculate_score(), obj.psqt);
-            }
+            // if board.calculate_score() != obj.psqt {
+            //     println!("assertion `{:?} == {:?}` failed", board.calculate_score(), obj.psqt);
+            // } else {
+            //     println!("assertion `{:?} == {:?}` success", board.calculate_score(), obj.psqt);
+            // }
 
-            // assert_eq!(board.calculate_score(), obj.psqt);
+            assert_eq!(board.calculate_score(), obj.psqt);
         }
     }
